Feature #98 - Refactor UI to use table widgets
Backend updated accordingly. Other various tweaks/fixes as needed
