--- Subsystems/cmdGui/CommandSystemDialog.py
@@ -1,1176 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file '/home/lbleier/cFS/tools/cFS-GroundSystem/Subsystems/cmdGui/CommandSystemDialog.ui'
-#
-# Created by: PyQt5 UI code generator 5.10.1
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-class Ui_CommandSystemDialog(object):
-    def setupUi(self, CommandSystemDialog):
-        CommandSystemDialog.setObjectName("CommandSystemDialog")
-        CommandSystemDialog.resize(772, 1020)
-        self.layoutWidget_2 = QtWidgets.QWidget(CommandSystemDialog)
-        self.layoutWidget_2.setGeometry(QtCore.QRect(80, 20, 621, 25))
-        self.layoutWidget_2.setObjectName("layoutWidget_2")
-        self.horizontalLayout = QtWidgets.QHBoxLayout(self.layoutWidget_2)
-        self.horizontalLayout.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout.setObjectName("horizontalLayout")
-        spacerItem = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem)
-        self.label = QtWidgets.QLabel(self.layoutWidget_2)
-        self.label.setObjectName("label")
-        self.horizontalLayout.addWidget(self.label)
-        spacerItem1 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem1)
-        self.layoutWidget_3 = QtWidgets.QWidget(CommandSystemDialog)
-        self.layoutWidget_3.setGeometry(QtCore.QRect(80, 60, 621, 33))
-        self.layoutWidget_3.setObjectName("layoutWidget_3")
-        self.horizontalLayout_2 = QtWidgets.QHBoxLayout(self.layoutWidget_3)
-        self.horizontalLayout_2.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_2.setObjectName("horizontalLayout_2")
-        spacerItem2 = QtWidgets.QSpacerItem(90, 20, QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_2.addItem(spacerItem2)
-        spacerItem3 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_2.addItem(spacerItem3)
-        self.label_2 = QtWidgets.QLabel(self.layoutWidget_3)
-        self.label_2.setObjectName("label_2")
-        self.horizontalLayout_2.addWidget(self.label_2)
-        spacerItem4 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_2.addItem(spacerItem4)
-        self.buttonBox = QtWidgets.QDialogButtonBox(self.layoutWidget_3)
-        self.buttonBox.setOrientation(QtCore.Qt.Vertical)
-        self.buttonBox.setStandardButtons(QtWidgets.QDialogButtonBox.Close)
-        self.buttonBox.setCenterButtons(True)
-        self.buttonBox.setObjectName("buttonBox")
-        self.horizontalLayout_2.addWidget(self.buttonBox)
-        self.layoutWidget = QtWidgets.QWidget(CommandSystemDialog)
-        self.layoutWidget.setGeometry(QtCore.QRect(70, 120, 630, 857))
-        self.layoutWidget.setObjectName("layoutWidget")
-        self.gridLayout = QtWidgets.QGridLayout(self.layoutWidget)
-        self.gridLayout.setContentsMargins(0, 0, 0, 0)
-        self.gridLayout.setObjectName("gridLayout")
-        self.quickButton_3 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_3.sizePolicy().hasHeightForWidth())
-        self.quickButton_3.setSizePolicy(sizePolicy)
-        self.quickButton_3.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_3.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_3.setText("")
-        self.quickButton_3.setObjectName("quickButton_3")
-        self.gridLayout.addWidget(self.quickButton_3, 3, 4, 1, 1)
-        self.SubsysBrowser_6 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_6.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_6.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_6.setReadOnly(True)
-        self.SubsysBrowser_6.setObjectName("SubsysBrowser_6")
-        self.gridLayout.addWidget(self.SubsysBrowser_6, 7, 0, 1, 1)
-        self.quickButton_4 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_4.sizePolicy().hasHeightForWidth())
-        self.quickButton_4.setSizePolicy(sizePolicy)
-        self.quickButton_4.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_4.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_4.setText("")
-        self.quickButton_4.setObjectName("quickButton_4")
-        self.gridLayout.addWidget(self.quickButton_4, 4, 4, 1, 1)
-        self.SubsysBrowser_18 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_18.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_18.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_18.setReadOnly(True)
-        self.SubsysBrowser_18.setObjectName("SubsysBrowser_18")
-        self.gridLayout.addWidget(self.SubsysBrowser_18, 19, 0, 1, 1)
-        self.lineEditPktId_18 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_18.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_18.setSizePolicy(sizePolicy)
-        self.lineEditPktId_18.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_18.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_18.setObjectName("lineEditPktId_18")
-        self.gridLayout.addWidget(self.lineEditPktId_18, 19, 1, 1, 1)
-        self.lineEdit_18 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_18.sizePolicy().hasHeightForWidth())
-        self.lineEdit_18.setSizePolicy(sizePolicy)
-        self.lineEdit_18.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_18.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_18.setObjectName("lineEdit_18")
-        self.gridLayout.addWidget(self.lineEdit_18, 19, 2, 1, 1)
-        self.pushButton_18 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_18.sizePolicy().hasHeightForWidth())
-        self.pushButton_18.setSizePolicy(sizePolicy)
-        self.pushButton_18.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_18.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_18.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_18.setObjectName("pushButton_18")
-        self.gridLayout.addWidget(self.pushButton_18, 19, 3, 1, 1)
-        self.quickButton_19 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_19.sizePolicy().hasHeightForWidth())
-        self.quickButton_19.setSizePolicy(sizePolicy)
-        self.quickButton_19.setMinimumSize(QtCore.QSize(121, 32))
-        self.quickButton_19.setMaximumSize(QtCore.QSize(121, 32))
-        self.quickButton_19.setText("")
-        self.quickButton_19.setObjectName("quickButton_19")
-        self.gridLayout.addWidget(self.quickButton_19, 19, 4, 1, 1)
-        self.SubsysBrowser_19 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_19.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_19.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_19.setReadOnly(True)
-        self.SubsysBrowser_19.setObjectName("SubsysBrowser_19")
-        self.gridLayout.addWidget(self.SubsysBrowser_19, 20, 0, 1, 1)
-        self.lineEditPktId_19 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_19.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_19.setSizePolicy(sizePolicy)
-        self.lineEditPktId_19.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_19.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_19.setObjectName("lineEditPktId_19")
-        self.gridLayout.addWidget(self.lineEditPktId_19, 20, 1, 1, 1)
-        self.lineEdit_19 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_19.sizePolicy().hasHeightForWidth())
-        self.lineEdit_19.setSizePolicy(sizePolicy)
-        self.lineEdit_19.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_19.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_19.setObjectName("lineEdit_19")
-        self.gridLayout.addWidget(self.lineEdit_19, 20, 2, 1, 1)
-        self.pushButton_19 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_19.sizePolicy().hasHeightForWidth())
-        self.pushButton_19.setSizePolicy(sizePolicy)
-        self.pushButton_19.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_19.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_19.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_19.setObjectName("pushButton_19")
-        self.gridLayout.addWidget(self.pushButton_19, 20, 3, 1, 1)
-        self.SubsysBrowser_16 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_16.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_16.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_16.setReadOnly(True)
-        self.SubsysBrowser_16.setObjectName("SubsysBrowser_16")
-        self.gridLayout.addWidget(self.SubsysBrowser_16, 17, 0, 1, 1)
-        self.lineEditPktId_16 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_16.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_16.setSizePolicy(sizePolicy)
-        self.lineEditPktId_16.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_16.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_16.setObjectName("lineEditPktId_16")
-        self.gridLayout.addWidget(self.lineEditPktId_16, 17, 1, 1, 1)
-        self.lineEdit_16 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_16.sizePolicy().hasHeightForWidth())
-        self.lineEdit_16.setSizePolicy(sizePolicy)
-        self.lineEdit_16.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_16.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_16.setObjectName("lineEdit_16")
-        self.gridLayout.addWidget(self.lineEdit_16, 17, 2, 1, 1)
-        self.pushButton_16 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_16.sizePolicy().hasHeightForWidth())
-        self.pushButton_16.setSizePolicy(sizePolicy)
-        self.pushButton_16.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_16.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_16.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_16.setObjectName("pushButton_16")
-        self.gridLayout.addWidget(self.pushButton_16, 17, 3, 1, 1)
-        self.quickButton_17 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_17.sizePolicy().hasHeightForWidth())
-        self.quickButton_17.setSizePolicy(sizePolicy)
-        self.quickButton_17.setMinimumSize(QtCore.QSize(121, 32))
-        self.quickButton_17.setMaximumSize(QtCore.QSize(121, 32))
-        self.quickButton_17.setText("")
-        self.quickButton_17.setObjectName("quickButton_17")
-        self.gridLayout.addWidget(self.quickButton_17, 17, 4, 1, 1)
-        self.quickButton_9 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_9.sizePolicy().hasHeightForWidth())
-        self.quickButton_9.setSizePolicy(sizePolicy)
-        self.quickButton_9.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_9.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_9.setText("")
-        self.quickButton_9.setObjectName("quickButton_9")
-        self.gridLayout.addWidget(self.quickButton_9, 9, 4, 1, 1)
-        self.SubsysBrowser_9 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_9.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_9.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_9.setReadOnly(True)
-        self.SubsysBrowser_9.setObjectName("SubsysBrowser_9")
-        self.gridLayout.addWidget(self.SubsysBrowser_9, 10, 0, 1, 1)
-        self.SubsysBrowser_3 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_3.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_3.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_3.setReadOnly(True)
-        self.SubsysBrowser_3.setObjectName("SubsysBrowser_3")
-        self.gridLayout.addWidget(self.SubsysBrowser_3, 4, 0, 1, 1)
-        self.pushButton_5 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_5.sizePolicy().hasHeightForWidth())
-        self.pushButton_5.setSizePolicy(sizePolicy)
-        self.pushButton_5.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_5.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_5.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_5.setObjectName("pushButton_5")
-        self.gridLayout.addWidget(self.pushButton_5, 6, 3, 1, 1)
-        self.lineEditPktId_4 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_4.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_4.setSizePolicy(sizePolicy)
-        self.lineEditPktId_4.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_4.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_4.setObjectName("lineEditPktId_4")
-        self.gridLayout.addWidget(self.lineEditPktId_4, 5, 1, 1, 1)
-        self.label_4 = QtWidgets.QLabel(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_4.sizePolicy().hasHeightForWidth())
-        self.label_4.setSizePolicy(sizePolicy)
-        self.label_4.setMinimumSize(QtCore.QSize(81, 31))
-        self.label_4.setMaximumSize(QtCore.QSize(81, 31))
-        self.label_4.setObjectName("label_4")
-        self.gridLayout.addWidget(self.label_4, 0, 1, 1, 1)
-        self.SubsysBrowser_4 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_4.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_4.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_4.setReadOnly(True)
-        self.SubsysBrowser_4.setObjectName("SubsysBrowser_4")
-        self.gridLayout.addWidget(self.SubsysBrowser_4, 5, 0, 1, 1)
-        self.lineEdit_5 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_5.sizePolicy().hasHeightForWidth())
-        self.lineEdit_5.setSizePolicy(sizePolicy)
-        self.lineEdit_5.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_5.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_5.setObjectName("lineEdit_5")
-        self.gridLayout.addWidget(self.lineEdit_5, 6, 2, 1, 1)
-        self.lineEdit_4 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_4.sizePolicy().hasHeightForWidth())
-        self.lineEdit_4.setSizePolicy(sizePolicy)
-        self.lineEdit_4.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_4.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_4.setObjectName("lineEdit_4")
-        self.gridLayout.addWidget(self.lineEdit_4, 5, 2, 1, 1)
-        self.pushButton_3 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_3.sizePolicy().hasHeightForWidth())
-        self.pushButton_3.setSizePolicy(sizePolicy)
-        self.pushButton_3.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_3.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_3.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_3.setObjectName("pushButton_3")
-        self.gridLayout.addWidget(self.pushButton_3, 4, 3, 1, 1)
-        self.lineEdit_11 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_11.sizePolicy().hasHeightForWidth())
-        self.lineEdit_11.setSizePolicy(sizePolicy)
-        self.lineEdit_11.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_11.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_11.setObjectName("lineEdit_11")
-        self.gridLayout.addWidget(self.lineEdit_11, 12, 2, 1, 1)
-        self.pushButton_11 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_11.sizePolicy().hasHeightForWidth())
-        self.pushButton_11.setSizePolicy(sizePolicy)
-        self.pushButton_11.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_11.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_11.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_11.setObjectName("pushButton_11")
-        self.gridLayout.addWidget(self.pushButton_11, 12, 3, 1, 1)
-        self.quickButton_12 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_12.sizePolicy().hasHeightForWidth())
-        self.quickButton_12.setSizePolicy(sizePolicy)
-        self.quickButton_12.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_12.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_12.setText("")
-        self.quickButton_12.setObjectName("quickButton_12")
-        self.gridLayout.addWidget(self.quickButton_12, 12, 4, 1, 1)
-        self.SubsysBrowser_12 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_12.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_12.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_12.setReadOnly(True)
-        self.SubsysBrowser_12.setObjectName("SubsysBrowser_12")
-        self.gridLayout.addWidget(self.SubsysBrowser_12, 13, 0, 1, 1)
-        self.quickButton_2 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_2.sizePolicy().hasHeightForWidth())
-        self.quickButton_2.setSizePolicy(sizePolicy)
-        self.quickButton_2.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_2.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_2.setText("")
-        self.quickButton_2.setObjectName("quickButton_2")
-        self.gridLayout.addWidget(self.quickButton_2, 2, 4, 1, 1)
-        self.SubsysBrowser_2 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_2.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_2.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_2.setReadOnly(True)
-        self.SubsysBrowser_2.setObjectName("SubsysBrowser_2")
-        self.gridLayout.addWidget(self.SubsysBrowser_2, 3, 0, 1, 1)
-        self.SubsysBrowser_0 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_0.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_0.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_0.setReadOnly(True)
-        self.SubsysBrowser_0.setObjectName("SubsysBrowser_0")
-        self.gridLayout.addWidget(self.SubsysBrowser_0, 1, 0, 1, 1)
-        self.pushButton_10 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_10.sizePolicy().hasHeightForWidth())
-        self.pushButton_10.setSizePolicy(sizePolicy)
-        self.pushButton_10.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_10.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_10.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_10.setObjectName("pushButton_10")
-        self.gridLayout.addWidget(self.pushButton_10, 11, 3, 1, 1)
-        self.quickButton_11 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_11.sizePolicy().hasHeightForWidth())
-        self.quickButton_11.setSizePolicy(sizePolicy)
-        self.quickButton_11.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_11.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_11.setText("")
-        self.quickButton_11.setObjectName("quickButton_11")
-        self.gridLayout.addWidget(self.quickButton_11, 11, 4, 1, 1)
-        self.SubsysBrowser_11 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_11.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_11.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_11.setReadOnly(True)
-        self.SubsysBrowser_11.setObjectName("SubsysBrowser_11")
-        self.gridLayout.addWidget(self.SubsysBrowser_11, 12, 0, 1, 1)
-        self.lineEditPktId_11 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_11.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_11.setSizePolicy(sizePolicy)
-        self.lineEditPktId_11.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_11.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_11.setObjectName("lineEditPktId_11")
-        self.gridLayout.addWidget(self.lineEditPktId_11, 12, 1, 1, 1)
-        self.lineEditPktId_12 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_12.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_12.setSizePolicy(sizePolicy)
-        self.lineEditPktId_12.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_12.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_12.setObjectName("lineEditPktId_12")
-        self.gridLayout.addWidget(self.lineEditPktId_12, 13, 1, 1, 1)
-        self.lineEdit_0 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_0.sizePolicy().hasHeightForWidth())
-        self.lineEdit_0.setSizePolicy(sizePolicy)
-        self.lineEdit_0.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_0.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_0.setObjectName("lineEdit_0")
-        self.gridLayout.addWidget(self.lineEdit_0, 1, 2, 1, 1)
-        self.lineEditPktId_7 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_7.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_7.setSizePolicy(sizePolicy)
-        self.lineEditPktId_7.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_7.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_7.setObjectName("lineEditPktId_7")
-        self.gridLayout.addWidget(self.lineEditPktId_7, 8, 1, 1, 1)
-        self.lineEdit_7 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_7.sizePolicy().hasHeightForWidth())
-        self.lineEdit_7.setSizePolicy(sizePolicy)
-        self.lineEdit_7.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_7.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_7.setObjectName("lineEdit_7")
-        self.gridLayout.addWidget(self.lineEdit_7, 8, 2, 1, 1)
-        self.lineEditPktId_0 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_0.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_0.setSizePolicy(sizePolicy)
-        self.lineEditPktId_0.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_0.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_0.setObjectName("lineEditPktId_0")
-        self.gridLayout.addWidget(self.lineEditPktId_0, 1, 1, 1, 1)
-        self.pushButton_2 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_2.sizePolicy().hasHeightForWidth())
-        self.pushButton_2.setSizePolicy(sizePolicy)
-        self.pushButton_2.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_2.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_2.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_2.setObjectName("pushButton_2")
-        self.gridLayout.addWidget(self.pushButton_2, 3, 3, 1, 1)
-        self.label_5 = QtWidgets.QLabel(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_5.sizePolicy().hasHeightForWidth())
-        self.label_5.setSizePolicy(sizePolicy)
-        self.label_5.setMinimumSize(QtCore.QSize(81, 31))
-        self.label_5.setMaximumSize(QtCore.QSize(81, 31))
-        self.label_5.setObjectName("label_5")
-        self.gridLayout.addWidget(self.label_5, 0, 2, 1, 1)
-        self.pushButton_1 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_1.sizePolicy().hasHeightForWidth())
-        self.pushButton_1.setSizePolicy(sizePolicy)
-        self.pushButton_1.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_1.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_1.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_1.setObjectName("pushButton_1")
-        self.gridLayout.addWidget(self.pushButton_1, 2, 3, 1, 1)
-        self.SubsysBrowser_1 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_1.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_1.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_1.setReadOnly(True)
-        self.SubsysBrowser_1.setObjectName("SubsysBrowser_1")
-        self.gridLayout.addWidget(self.SubsysBrowser_1, 2, 0, 1, 1)
-        self.pushButton_7 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_7.sizePolicy().hasHeightForWidth())
-        self.pushButton_7.setSizePolicy(sizePolicy)
-        self.pushButton_7.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_7.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_7.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_7.setObjectName("pushButton_7")
-        self.gridLayout.addWidget(self.pushButton_7, 8, 3, 1, 1)
-        self.quickButton_8 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_8.sizePolicy().hasHeightForWidth())
-        self.quickButton_8.setSizePolicy(sizePolicy)
-        self.quickButton_8.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_8.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_8.setText("")
-        self.quickButton_8.setObjectName("quickButton_8")
-        self.gridLayout.addWidget(self.quickButton_8, 8, 4, 1, 1)
-        self.lineEditPktId_2 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_2.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_2.setSizePolicy(sizePolicy)
-        self.lineEditPktId_2.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_2.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_2.setObjectName("lineEditPktId_2")
-        self.gridLayout.addWidget(self.lineEditPktId_2, 3, 1, 1, 1)
-        self.lineEditPktId_5 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_5.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_5.setSizePolicy(sizePolicy)
-        self.lineEditPktId_5.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_5.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_5.setObjectName("lineEditPktId_5")
-        self.gridLayout.addWidget(self.lineEditPktId_5, 6, 1, 1, 1)
-        self.pushButton_6 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_6.sizePolicy().hasHeightForWidth())
-        self.pushButton_6.setSizePolicy(sizePolicy)
-        self.pushButton_6.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_6.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_6.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_6.setObjectName("pushButton_6")
-        self.gridLayout.addWidget(self.pushButton_6, 7, 3, 1, 1)
-        self.pushButton_9 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_9.sizePolicy().hasHeightForWidth())
-        self.pushButton_9.setSizePolicy(sizePolicy)
-        self.pushButton_9.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_9.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_9.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_9.setObjectName("pushButton_9")
-        self.gridLayout.addWidget(self.pushButton_9, 10, 3, 1, 1)
-        self.quickButton_10 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_10.sizePolicy().hasHeightForWidth())
-        self.quickButton_10.setSizePolicy(sizePolicy)
-        self.quickButton_10.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_10.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_10.setText("")
-        self.quickButton_10.setObjectName("quickButton_10")
-        self.gridLayout.addWidget(self.quickButton_10, 10, 4, 1, 1)
-        self.SubsysBrowser_15 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_15.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_15.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_15.setReadOnly(True)
-        self.SubsysBrowser_15.setObjectName("SubsysBrowser_15")
-        self.gridLayout.addWidget(self.SubsysBrowser_15, 16, 0, 1, 1)
-        self.lineEditPktId_15 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_15.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_15.setSizePolicy(sizePolicy)
-        self.lineEditPktId_15.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_15.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_15.setObjectName("lineEditPktId_15")
-        self.gridLayout.addWidget(self.lineEditPktId_15, 16, 1, 1, 1)
-        self.lineEdit_15 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_15.sizePolicy().hasHeightForWidth())
-        self.lineEdit_15.setSizePolicy(sizePolicy)
-        self.lineEdit_15.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_15.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_15.setObjectName("lineEdit_15")
-        self.gridLayout.addWidget(self.lineEdit_15, 16, 2, 1, 1)
-        self.pushButton_15 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_15.sizePolicy().hasHeightForWidth())
-        self.pushButton_15.setSizePolicy(sizePolicy)
-        self.pushButton_15.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_15.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_15.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_15.setObjectName("pushButton_15")
-        self.gridLayout.addWidget(self.pushButton_15, 16, 3, 1, 1)
-        self.quickButton_16 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_16.sizePolicy().hasHeightForWidth())
-        self.quickButton_16.setSizePolicy(sizePolicy)
-        self.quickButton_16.setMinimumSize(QtCore.QSize(121, 32))
-        self.quickButton_16.setMaximumSize(QtCore.QSize(121, 32))
-        self.quickButton_16.setText("")
-        self.quickButton_16.setObjectName("quickButton_16")
-        self.gridLayout.addWidget(self.quickButton_16, 16, 4, 1, 1)
-        self.quickButton_6 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_6.sizePolicy().hasHeightForWidth())
-        self.quickButton_6.setSizePolicy(sizePolicy)
-        self.quickButton_6.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_6.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_6.setText("")
-        self.quickButton_6.setObjectName("quickButton_6")
-        self.gridLayout.addWidget(self.quickButton_6, 6, 4, 1, 1)
-        self.lineEdit_12 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_12.sizePolicy().hasHeightForWidth())
-        self.lineEdit_12.setSizePolicy(sizePolicy)
-        self.lineEdit_12.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_12.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_12.setObjectName("lineEdit_12")
-        self.gridLayout.addWidget(self.lineEdit_12, 13, 2, 1, 1)
-        self.pushButton_12 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_12.sizePolicy().hasHeightForWidth())
-        self.pushButton_12.setSizePolicy(sizePolicy)
-        self.pushButton_12.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_12.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_12.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_12.setObjectName("pushButton_12")
-        self.gridLayout.addWidget(self.pushButton_12, 13, 3, 1, 1)
-        self.quickButton_13 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_13.sizePolicy().hasHeightForWidth())
-        self.quickButton_13.setSizePolicy(sizePolicy)
-        self.quickButton_13.setMinimumSize(QtCore.QSize(121, 32))
-        self.quickButton_13.setMaximumSize(QtCore.QSize(121, 32))
-        self.quickButton_13.setText("")
-        self.quickButton_13.setObjectName("quickButton_13")
-        self.gridLayout.addWidget(self.quickButton_13, 13, 4, 1, 1)
-        self.SubsysBrowser_13 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_13.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_13.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_13.setReadOnly(True)
-        self.SubsysBrowser_13.setObjectName("SubsysBrowser_13")
-        self.gridLayout.addWidget(self.SubsysBrowser_13, 14, 0, 1, 1)
-        self.SubsysBrowser_14 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_14.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_14.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_14.setReadOnly(True)
-        self.SubsysBrowser_14.setObjectName("SubsysBrowser_14")
-        self.gridLayout.addWidget(self.SubsysBrowser_14, 15, 0, 1, 1)
-        self.lineEditPktId_14 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_14.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_14.setSizePolicy(sizePolicy)
-        self.lineEditPktId_14.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_14.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_14.setObjectName("lineEditPktId_14")
-        self.gridLayout.addWidget(self.lineEditPktId_14, 15, 1, 1, 1)
-        self.lineEdit_14 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_14.sizePolicy().hasHeightForWidth())
-        self.lineEdit_14.setSizePolicy(sizePolicy)
-        self.lineEdit_14.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_14.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_14.setObjectName("lineEdit_14")
-        self.gridLayout.addWidget(self.lineEdit_14, 15, 2, 1, 1)
-        self.pushButton_14 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_14.sizePolicy().hasHeightForWidth())
-        self.pushButton_14.setSizePolicy(sizePolicy)
-        self.pushButton_14.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_14.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_14.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_14.setObjectName("pushButton_14")
-        self.gridLayout.addWidget(self.pushButton_14, 15, 3, 1, 1)
-        self.quickButton_15 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_15.sizePolicy().hasHeightForWidth())
-        self.quickButton_15.setSizePolicy(sizePolicy)
-        self.quickButton_15.setMinimumSize(QtCore.QSize(121, 32))
-        self.quickButton_15.setMaximumSize(QtCore.QSize(121, 32))
-        self.quickButton_15.setText("")
-        self.quickButton_15.setObjectName("quickButton_15")
-        self.gridLayout.addWidget(self.quickButton_15, 15, 4, 1, 1)
-        self.SubsysBrowser_8 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_8.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_8.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_8.setReadOnly(True)
-        self.SubsysBrowser_8.setObjectName("SubsysBrowser_8")
-        self.gridLayout.addWidget(self.SubsysBrowser_8, 9, 0, 1, 1)
-        self.lineEditPktId_8 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_8.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_8.setSizePolicy(sizePolicy)
-        self.lineEditPktId_8.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_8.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_8.setObjectName("lineEditPktId_8")
-        self.gridLayout.addWidget(self.lineEditPktId_8, 9, 1, 1, 1)
-        self.lineEditPktId_13 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_13.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_13.setSizePolicy(sizePolicy)
-        self.lineEditPktId_13.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_13.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_13.setObjectName("lineEditPktId_13")
-        self.gridLayout.addWidget(self.lineEditPktId_13, 14, 1, 1, 1)
-        self.lineEdit_13 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_13.sizePolicy().hasHeightForWidth())
-        self.lineEdit_13.setSizePolicy(sizePolicy)
-        self.lineEdit_13.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_13.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_13.setObjectName("lineEdit_13")
-        self.gridLayout.addWidget(self.lineEdit_13, 14, 2, 1, 1)
-        self.pushButton_13 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_13.sizePolicy().hasHeightForWidth())
-        self.pushButton_13.setSizePolicy(sizePolicy)
-        self.pushButton_13.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_13.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_13.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_13.setObjectName("pushButton_13")
-        self.gridLayout.addWidget(self.pushButton_13, 14, 3, 1, 1)
-        self.quickButton_14 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_14.sizePolicy().hasHeightForWidth())
-        self.quickButton_14.setSizePolicy(sizePolicy)
-        self.quickButton_14.setMinimumSize(QtCore.QSize(121, 32))
-        self.quickButton_14.setMaximumSize(QtCore.QSize(121, 32))
-        self.quickButton_14.setText("")
-        self.quickButton_14.setObjectName("quickButton_14")
-        self.gridLayout.addWidget(self.quickButton_14, 14, 4, 1, 1)
-        self.pushButton_4 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_4.sizePolicy().hasHeightForWidth())
-        self.pushButton_4.setSizePolicy(sizePolicy)
-        self.pushButton_4.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_4.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_4.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_4.setObjectName("pushButton_4")
-        self.gridLayout.addWidget(self.pushButton_4, 5, 3, 1, 1)
-        self.lineEdit_3 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_3.sizePolicy().hasHeightForWidth())
-        self.lineEdit_3.setSizePolicy(sizePolicy)
-        self.lineEdit_3.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_3.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_3.setObjectName("lineEdit_3")
-        self.gridLayout.addWidget(self.lineEdit_3, 4, 2, 1, 1)
-        self.SubsysBrowser_7 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_7.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_7.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_7.setReadOnly(True)
-        self.SubsysBrowser_7.setObjectName("SubsysBrowser_7")
-        self.gridLayout.addWidget(self.SubsysBrowser_7, 8, 0, 1, 1)
-        self.pushButton_0 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_0.sizePolicy().hasHeightForWidth())
-        self.pushButton_0.setSizePolicy(sizePolicy)
-        self.pushButton_0.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_0.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_0.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_0.setObjectName("pushButton_0")
-        self.gridLayout.addWidget(self.pushButton_0, 1, 3, 1, 1)
-        self.lineEditPktId_6 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_6.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_6.setSizePolicy(sizePolicy)
-        self.lineEditPktId_6.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_6.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_6.setObjectName("lineEditPktId_6")
-        self.gridLayout.addWidget(self.lineEditPktId_6, 7, 1, 1, 1)
-        self.SubsysBrowser_10 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_10.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_10.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_10.setReadOnly(True)
-        self.SubsysBrowser_10.setObjectName("SubsysBrowser_10")
-        self.gridLayout.addWidget(self.SubsysBrowser_10, 11, 0, 1, 1)
-        self.lineEditPktId_10 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_10.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_10.setSizePolicy(sizePolicy)
-        self.lineEditPktId_10.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_10.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_10.setObjectName("lineEditPktId_10")
-        self.gridLayout.addWidget(self.lineEditPktId_10, 11, 1, 1, 1)
-        self.lineEdit_10 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_10.sizePolicy().hasHeightForWidth())
-        self.lineEdit_10.setSizePolicy(sizePolicy)
-        self.lineEdit_10.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_10.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_10.setObjectName("lineEdit_10")
-        self.gridLayout.addWidget(self.lineEdit_10, 11, 2, 1, 1)
-        self.lineEdit_2 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_2.sizePolicy().hasHeightForWidth())
-        self.lineEdit_2.setSizePolicy(sizePolicy)
-        self.lineEdit_2.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_2.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_2.setObjectName("lineEdit_2")
-        self.gridLayout.addWidget(self.lineEdit_2, 3, 2, 1, 1)
-        self.SubsysBrowser_5 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_5.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_5.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_5.setReadOnly(True)
-        self.SubsysBrowser_5.setObjectName("SubsysBrowser_5")
-        self.gridLayout.addWidget(self.SubsysBrowser_5, 6, 0, 1, 1)
-        self.lineEditPktId_3 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_3.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_3.setSizePolicy(sizePolicy)
-        self.lineEditPktId_3.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_3.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_3.setObjectName("lineEditPktId_3")
-        self.gridLayout.addWidget(self.lineEditPktId_3, 4, 1, 1, 1)
-        self.lineEdit_1 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_1.sizePolicy().hasHeightForWidth())
-        self.lineEdit_1.setSizePolicy(sizePolicy)
-        self.lineEdit_1.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_1.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_1.setObjectName("lineEdit_1")
-        self.gridLayout.addWidget(self.lineEdit_1, 2, 2, 1, 1)
-        self.lineEdit_8 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_8.sizePolicy().hasHeightForWidth())
-        self.lineEdit_8.setSizePolicy(sizePolicy)
-        self.lineEdit_8.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_8.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_8.setObjectName("lineEdit_8")
-        self.gridLayout.addWidget(self.lineEdit_8, 9, 2, 1, 1)
-        self.pushButton_8 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_8.sizePolicy().hasHeightForWidth())
-        self.pushButton_8.setSizePolicy(sizePolicy)
-        self.pushButton_8.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_8.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_8.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_8.setObjectName("pushButton_8")
-        self.gridLayout.addWidget(self.pushButton_8, 9, 3, 1, 1)
-        self.lineEditPktId_1 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_1.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_1.setSizePolicy(sizePolicy)
-        self.lineEditPktId_1.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_1.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_1.setObjectName("lineEditPktId_1")
-        self.gridLayout.addWidget(self.lineEditPktId_1, 2, 1, 1, 1)
-        self.lineEdit_6 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_6.sizePolicy().hasHeightForWidth())
-        self.lineEdit_6.setSizePolicy(sizePolicy)
-        self.lineEdit_6.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_6.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_6.setObjectName("lineEdit_6")
-        self.gridLayout.addWidget(self.lineEdit_6, 7, 2, 1, 1)
-        self.lineEditPktId_9 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_9.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_9.setSizePolicy(sizePolicy)
-        self.lineEditPktId_9.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_9.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_9.setObjectName("lineEditPktId_9")
-        self.gridLayout.addWidget(self.lineEditPktId_9, 10, 1, 1, 1)
-        self.lineEdit_9 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_9.sizePolicy().hasHeightForWidth())
-        self.lineEdit_9.setSizePolicy(sizePolicy)
-        self.lineEdit_9.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_9.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_9.setObjectName("lineEdit_9")
-        self.gridLayout.addWidget(self.lineEdit_9, 10, 2, 1, 1)
-        self.quickButton_7 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_7.sizePolicy().hasHeightForWidth())
-        self.quickButton_7.setSizePolicy(sizePolicy)
-        self.quickButton_7.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_7.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_7.setText("")
-        self.quickButton_7.setObjectName("quickButton_7")
-        self.gridLayout.addWidget(self.quickButton_7, 7, 4, 1, 1)
-        self.quickButton_5 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_5.sizePolicy().hasHeightForWidth())
-        self.quickButton_5.setSizePolicy(sizePolicy)
-        self.quickButton_5.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_5.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_5.setText("")
-        self.quickButton_5.setObjectName("quickButton_5")
-        self.gridLayout.addWidget(self.quickButton_5, 5, 4, 1, 1)
-        self.SubsysBrowser_17 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_17.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_17.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_17.setReadOnly(True)
-        self.SubsysBrowser_17.setObjectName("SubsysBrowser_17")
-        self.gridLayout.addWidget(self.SubsysBrowser_17, 18, 0, 1, 1)
-        self.lineEditPktId_17 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_17.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_17.setSizePolicy(sizePolicy)
-        self.lineEditPktId_17.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_17.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_17.setObjectName("lineEditPktId_17")
-        self.gridLayout.addWidget(self.lineEditPktId_17, 18, 1, 1, 1)
-        self.lineEdit_17 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_17.sizePolicy().hasHeightForWidth())
-        self.lineEdit_17.setSizePolicy(sizePolicy)
-        self.lineEdit_17.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_17.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_17.setObjectName("lineEdit_17")
-        self.gridLayout.addWidget(self.lineEdit_17, 18, 2, 1, 1)
-        self.pushButton_17 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_17.sizePolicy().hasHeightForWidth())
-        self.pushButton_17.setSizePolicy(sizePolicy)
-        self.pushButton_17.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_17.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_17.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_17.setObjectName("pushButton_17")
-        self.gridLayout.addWidget(self.pushButton_17, 18, 3, 1, 1)
-        self.quickButton_18 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_18.sizePolicy().hasHeightForWidth())
-        self.quickButton_18.setSizePolicy(sizePolicy)
-        self.quickButton_18.setMinimumSize(QtCore.QSize(121, 32))
-        self.quickButton_18.setMaximumSize(QtCore.QSize(121, 32))
-        self.quickButton_18.setText("")
-        self.quickButton_18.setObjectName("quickButton_18")
-        self.gridLayout.addWidget(self.quickButton_18, 18, 4, 1, 1)
-        self.quickButton_20 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_20.sizePolicy().hasHeightForWidth())
-        self.quickButton_20.setSizePolicy(sizePolicy)
-        self.quickButton_20.setMinimumSize(QtCore.QSize(121, 32))
-        self.quickButton_20.setMaximumSize(QtCore.QSize(121, 32))
-        self.quickButton_20.setText("")
-        self.quickButton_20.setObjectName("quickButton_20")
-        self.gridLayout.addWidget(self.quickButton_20, 20, 4, 1, 1)
-        self.SubsysBrowser_20 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_20.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_20.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_20.setReadOnly(True)
-        self.SubsysBrowser_20.setObjectName("SubsysBrowser_20")
-        self.gridLayout.addWidget(self.SubsysBrowser_20, 21, 0, 1, 1)
-        self.lineEditPktId_20 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_20.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_20.setSizePolicy(sizePolicy)
-        self.lineEditPktId_20.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_20.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_20.setObjectName("lineEditPktId_20")
-        self.gridLayout.addWidget(self.lineEditPktId_20, 21, 1, 1, 1)
-        self.lineEdit_20 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_20.sizePolicy().hasHeightForWidth())
-        self.lineEdit_20.setSizePolicy(sizePolicy)
-        self.lineEdit_20.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_20.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_20.setObjectName("lineEdit_20")
-        self.gridLayout.addWidget(self.lineEdit_20, 21, 2, 1, 1)
-        self.pushButton_20 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_20.sizePolicy().hasHeightForWidth())
-        self.pushButton_20.setSizePolicy(sizePolicy)
-        self.pushButton_20.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_20.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_20.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_20.setObjectName("pushButton_20")
-        self.gridLayout.addWidget(self.pushButton_20, 21, 3, 1, 1)
-        self.quickButton_21 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_21.sizePolicy().hasHeightForWidth())
-        self.quickButton_21.setSizePolicy(sizePolicy)
-        self.quickButton_21.setMinimumSize(QtCore.QSize(121, 32))
-        self.quickButton_21.setMaximumSize(QtCore.QSize(121, 32))
-        self.quickButton_21.setText("")
-        self.quickButton_21.setObjectName("quickButton_21")
-        self.gridLayout.addWidget(self.quickButton_21, 21, 4, 1, 1)
-        self.SubsysBrowser_21 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.SubsysBrowser_21.setMinimumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_21.setMaximumSize(QtCore.QSize(201, 31))
-        self.SubsysBrowser_21.setReadOnly(True)
-        self.SubsysBrowser_21.setObjectName("SubsysBrowser_21")
-        self.gridLayout.addWidget(self.SubsysBrowser_21, 22, 0, 1, 1)
-        self.lineEditPktId_21 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEditPktId_21.sizePolicy().hasHeightForWidth())
-        self.lineEditPktId_21.setSizePolicy(sizePolicy)
-        self.lineEditPktId_21.setMinimumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_21.setMaximumSize(QtCore.QSize(81, 31))
-        self.lineEditPktId_21.setObjectName("lineEditPktId_21")
-        self.gridLayout.addWidget(self.lineEditPktId_21, 22, 1, 1, 1)
-        self.lineEdit_21 = QtWidgets.QLineEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.lineEdit_21.sizePolicy().hasHeightForWidth())
-        self.lineEdit_21.setSizePolicy(sizePolicy)
-        self.lineEdit_21.setMinimumSize(QtCore.QSize(100, 31))
-        self.lineEdit_21.setMaximumSize(QtCore.QSize(100, 31))
-        self.lineEdit_21.setObjectName("lineEdit_21")
-        self.gridLayout.addWidget(self.lineEdit_21, 22, 2, 1, 1)
-        self.pushButton_21 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_21.sizePolicy().hasHeightForWidth())
-        self.pushButton_21.setSizePolicy(sizePolicy)
-        self.pushButton_21.setMinimumSize(QtCore.QSize(101, 31))
-        self.pushButton_21.setMaximumSize(QtCore.QSize(101, 31))
-        self.pushButton_21.setFocusPolicy(QtCore.Qt.NoFocus)
-        self.pushButton_21.setObjectName("pushButton_21")
-        self.gridLayout.addWidget(self.pushButton_21, 22, 3, 1, 1)
-        self.quickButton_22 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_22.sizePolicy().hasHeightForWidth())
-        self.quickButton_22.setSizePolicy(sizePolicy)
-        self.quickButton_22.setMinimumSize(QtCore.QSize(121, 32))
-        self.quickButton_22.setMaximumSize(QtCore.QSize(121, 32))
-        self.quickButton_22.setText("")
-        self.quickButton_22.setObjectName("quickButton_22")
-        self.gridLayout.addWidget(self.quickButton_22, 22, 4, 1, 1)
-        self.quickButton_1 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.quickButton_1.sizePolicy().hasHeightForWidth())
-        self.quickButton_1.setSizePolicy(sizePolicy)
-        self.quickButton_1.setMinimumSize(QtCore.QSize(121, 0))
-        self.quickButton_1.setMaximumSize(QtCore.QSize(121, 16777215))
-        self.quickButton_1.setText("")
-        self.quickButton_1.setObjectName("quickButton_1")
-        self.gridLayout.addWidget(self.quickButton_1, 1, 4, 1, 1)
-        self.label_3 = QtWidgets.QLabel(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_3.sizePolicy().hasHeightForWidth())
-        self.label_3.setSizePolicy(sizePolicy)
-        self.label_3.setMinimumSize(QtCore.QSize(201, 31))
-        self.label_3.setMaximumSize(QtCore.QSize(201, 31))
-        self.label_3.setObjectName("label_3")
-        self.gridLayout.addWidget(self.label_3, 0, 0, 1, 1)
-
-        self.retranslateUi(CommandSystemDialog)
-        self.buttonBox.clicked['QAbstractButton*'].connect(CommandSystemDialog.close)
-        QtCore.QMetaObject.connectSlotsByName(CommandSystemDialog)
-
-    def retranslateUi(self, CommandSystemDialog):
-        _translate = QtCore.QCoreApplication.translate
-        CommandSystemDialog.setWindowTitle(_translate("CommandSystemDialog", "Command System Main Page"))
-        self.label.setText(_translate("CommandSystemDialog", "cFE/CFS Subsystem Commands"))
-        self.label_2.setText(_translate("CommandSystemDialog", "Available Pages"))
-        self.pushButton_18.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_19.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_16.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_5.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.label_4.setText(_translate("CommandSystemDialog", "Packet ID"))
-        self.pushButton_3.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_11.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_10.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_2.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.label_5.setText(_translate("CommandSystemDialog", "Send To"))
-        self.pushButton_1.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_7.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_6.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_9.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_15.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_12.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_14.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_13.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_4.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_0.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_8.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_17.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_20.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.pushButton_21.setText(_translate("CommandSystemDialog", "Display Page"))
-        self.label_3.setText(_translate("CommandSystemDialog", "Subsystem/Page"))
-
