--- Subsystems/cmdGui/GenericCommandDialog.py
@@ -1,603 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file '/home/lbleier/cFS/tools/cFS-GroundSystem/Subsystems/cmdGui/GenericCommandDialog.ui'
-#
-# Created by: PyQt5 UI code generator 5.10.1
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-class Ui_GenericCommandDialog(object):
-    def setupUi(self, GenericCommandDialog):
-        GenericCommandDialog.setObjectName("GenericCommandDialog")
-        GenericCommandDialog.resize(549, 1058)
-        self.verticalLayout_6 = QtWidgets.QVBoxLayout(GenericCommandDialog)
-        self.verticalLayout_6.setObjectName("verticalLayout_6")
-        self.verticalLayout_4 = QtWidgets.QVBoxLayout()
-        self.verticalLayout_4.setObjectName("verticalLayout_4")
-        self.horizontalLayout = QtWidgets.QHBoxLayout()
-        self.horizontalLayout.setObjectName("horizontalLayout")
-        self.verticalLayout = QtWidgets.QVBoxLayout()
-        self.verticalLayout.setObjectName("verticalLayout")
-        self.subSystemCommandPageLabel = QtWidgets.QLabel(GenericCommandDialog)
-        self.subSystemCommandPageLabel.setAlignment(QtCore.Qt.AlignCenter)
-        self.subSystemCommandPageLabel.setObjectName("subSystemCommandPageLabel")
-        self.verticalLayout.addWidget(self.subSystemCommandPageLabel)
-        self.subSystemLineEdit = QtWidgets.QLineEdit(GenericCommandDialog)
-        self.subSystemLineEdit.setMinimumSize(QtCore.QSize(159, 31))
-        self.subSystemLineEdit.setMaximumSize(QtCore.QSize(300, 31))
-        self.subSystemLineEdit.setObjectName("subSystemLineEdit")
-        self.verticalLayout.addWidget(self.subSystemLineEdit)
-        self.horizontalLayout.addLayout(self.verticalLayout)
-        self.verticalLayout_2 = QtWidgets.QVBoxLayout()
-        self.verticalLayout_2.setObjectName("verticalLayout_2")
-        self.packetIdLabel = QtWidgets.QLabel(GenericCommandDialog)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.packetIdLabel.sizePolicy().hasHeightForWidth())
-        self.packetIdLabel.setSizePolicy(sizePolicy)
-        self.packetIdLabel.setMinimumSize(QtCore.QSize(0, 13))
-        self.packetIdLabel.setMaximumSize(QtCore.QSize(16777193, 13))
-        self.packetIdLabel.setObjectName("packetIdLabel")
-        self.verticalLayout_2.addWidget(self.packetIdLabel)
-        self.packetId = QtWidgets.QLCDNumber(GenericCommandDialog)
-        self.packetId.setObjectName("packetId")
-        self.verticalLayout_2.addWidget(self.packetId)
-        self.horizontalLayout.addLayout(self.verticalLayout_2)
-        self.verticalLayout_3 = QtWidgets.QVBoxLayout()
-        self.verticalLayout_3.setObjectName("verticalLayout_3")
-        self.label_5 = QtWidgets.QLabel(GenericCommandDialog)
-        self.label_5.setAlignment(QtCore.Qt.AlignCenter)
-        self.label_5.setObjectName("label_5")
-        self.verticalLayout_3.addWidget(self.label_5)
-        self.commandAddressLineEdit = QtWidgets.QLineEdit(GenericCommandDialog)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.commandAddressLineEdit.sizePolicy().hasHeightForWidth())
-        self.commandAddressLineEdit.setSizePolicy(sizePolicy)
-        self.commandAddressLineEdit.setMinimumSize(QtCore.QSize(135, 31))
-        self.commandAddressLineEdit.setMaximumSize(QtCore.QSize(135, 31))
-        self.commandAddressLineEdit.setObjectName("commandAddressLineEdit")
-        self.verticalLayout_3.addWidget(self.commandAddressLineEdit)
-        self.horizontalLayout.addLayout(self.verticalLayout_3)
-        spacerItem = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem)
-        self.buttonBox = QtWidgets.QDialogButtonBox(GenericCommandDialog)
-        self.buttonBox.setOrientation(QtCore.Qt.Horizontal)
-        self.buttonBox.setStandardButtons(QtWidgets.QDialogButtonBox.Close)
-        self.buttonBox.setObjectName("buttonBox")
-        self.horizontalLayout.addWidget(self.buttonBox)
-        self.verticalLayout_4.addLayout(self.horizontalLayout)
-        self.label = QtWidgets.QLabel(GenericCommandDialog)
-        self.label.setAlignment(QtCore.Qt.AlignCenter)
-        self.label.setObjectName("label")
-        self.verticalLayout_4.addWidget(self.label)
-        self.verticalLayout_6.addLayout(self.verticalLayout_4)
-        self.scrollArea = QtWidgets.QScrollArea(GenericCommandDialog)
-        self.scrollArea.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOn)
-        self.scrollArea.setWidgetResizable(True)
-        self.scrollArea.setObjectName("scrollArea")
-        self.scrollAreaWidgetContents = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents.setGeometry(QtCore.QRect(0, 0, 515, 1000))
-        self.scrollAreaWidgetContents.setMinimumSize(QtCore.QSize(0, 1000))
-        self.scrollAreaWidgetContents.setObjectName("scrollAreaWidgetContents")
-        self.verticalLayout_5 = QtWidgets.QVBoxLayout(self.scrollAreaWidgetContents)
-        self.verticalLayout_5.setObjectName("verticalLayout_5")
-        self.horizontalLayout_2 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_2.setObjectName("horizontalLayout_2")
-        self.itemLabelTextBrowser_1 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_1.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_1.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_1.setReadOnly(True)
-        self.itemLabelTextBrowser_1.setObjectName("itemLabelTextBrowser_1")
-        self.horizontalLayout_2.addWidget(self.itemLabelTextBrowser_1)
-        self.SendButton_1 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_1.sizePolicy().hasHeightForWidth())
-        self.SendButton_1.setSizePolicy(sizePolicy)
-        self.SendButton_1.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_1.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_1.setObjectName("SendButton_1")
-        self.horizontalLayout_2.addWidget(self.SendButton_1)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_2)
-        self.horizontalLayout_3 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_3.setObjectName("horizontalLayout_3")
-        self.itemLabelTextBrowser_2 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_2.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_2.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_2.setReadOnly(True)
-        self.itemLabelTextBrowser_2.setObjectName("itemLabelTextBrowser_2")
-        self.horizontalLayout_3.addWidget(self.itemLabelTextBrowser_2)
-        self.SendButton_2 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_2.sizePolicy().hasHeightForWidth())
-        self.SendButton_2.setSizePolicy(sizePolicy)
-        self.SendButton_2.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_2.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_2.setObjectName("SendButton_2")
-        self.horizontalLayout_3.addWidget(self.SendButton_2)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_3)
-        self.horizontalLayout_4 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_4.setObjectName("horizontalLayout_4")
-        self.itemLabelTextBrowser_3 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_3.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_3.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_3.setReadOnly(True)
-        self.itemLabelTextBrowser_3.setObjectName("itemLabelTextBrowser_3")
-        self.horizontalLayout_4.addWidget(self.itemLabelTextBrowser_3)
-        self.SendButton_3 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_3.sizePolicy().hasHeightForWidth())
-        self.SendButton_3.setSizePolicy(sizePolicy)
-        self.SendButton_3.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_3.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_3.setObjectName("SendButton_3")
-        self.horizontalLayout_4.addWidget(self.SendButton_3)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_4)
-        self.horizontalLayout_5 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_5.setObjectName("horizontalLayout_5")
-        self.itemLabelTextBrowser_4 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_4.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_4.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_4.setReadOnly(True)
-        self.itemLabelTextBrowser_4.setObjectName("itemLabelTextBrowser_4")
-        self.horizontalLayout_5.addWidget(self.itemLabelTextBrowser_4)
-        self.SendButton_4 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_4.sizePolicy().hasHeightForWidth())
-        self.SendButton_4.setSizePolicy(sizePolicy)
-        self.SendButton_4.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_4.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_4.setObjectName("SendButton_4")
-        self.horizontalLayout_5.addWidget(self.SendButton_4)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_5)
-        self.horizontalLayout_6 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_6.setObjectName("horizontalLayout_6")
-        self.itemLabelTextBrowser_5 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_5.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_5.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_5.setReadOnly(True)
-        self.itemLabelTextBrowser_5.setObjectName("itemLabelTextBrowser_5")
-        self.horizontalLayout_6.addWidget(self.itemLabelTextBrowser_5)
-        self.SendButton_5 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_5.sizePolicy().hasHeightForWidth())
-        self.SendButton_5.setSizePolicy(sizePolicy)
-        self.SendButton_5.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_5.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_5.setObjectName("SendButton_5")
-        self.horizontalLayout_6.addWidget(self.SendButton_5)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_6)
-        self.horizontalLayout_7 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_7.setObjectName("horizontalLayout_7")
-        self.itemLabelTextBrowser_6 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_6.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_6.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_6.setReadOnly(True)
-        self.itemLabelTextBrowser_6.setObjectName("itemLabelTextBrowser_6")
-        self.horizontalLayout_7.addWidget(self.itemLabelTextBrowser_6)
-        self.SendButton_6 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_6.sizePolicy().hasHeightForWidth())
-        self.SendButton_6.setSizePolicy(sizePolicy)
-        self.SendButton_6.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_6.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_6.setObjectName("SendButton_6")
-        self.horizontalLayout_7.addWidget(self.SendButton_6)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_7)
-        self.horizontalLayout_8 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_8.setObjectName("horizontalLayout_8")
-        self.itemLabelTextBrowser_7 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_7.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_7.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_7.setReadOnly(True)
-        self.itemLabelTextBrowser_7.setObjectName("itemLabelTextBrowser_7")
-        self.horizontalLayout_8.addWidget(self.itemLabelTextBrowser_7)
-        self.SendButton_7 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_7.sizePolicy().hasHeightForWidth())
-        self.SendButton_7.setSizePolicy(sizePolicy)
-        self.SendButton_7.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_7.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_7.setObjectName("SendButton_7")
-        self.horizontalLayout_8.addWidget(self.SendButton_7)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_8)
-        self.horizontalLayout_9 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_9.setObjectName("horizontalLayout_9")
-        self.itemLabelTextBrowser_8 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_8.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_8.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_8.setReadOnly(True)
-        self.itemLabelTextBrowser_8.setObjectName("itemLabelTextBrowser_8")
-        self.horizontalLayout_9.addWidget(self.itemLabelTextBrowser_8)
-        self.SendButton_8 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_8.sizePolicy().hasHeightForWidth())
-        self.SendButton_8.setSizePolicy(sizePolicy)
-        self.SendButton_8.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_8.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_8.setObjectName("SendButton_8")
-        self.horizontalLayout_9.addWidget(self.SendButton_8)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_9)
-        self.horizontalLayout_10 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_10.setObjectName("horizontalLayout_10")
-        self.itemLabelTextBrowser_9 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_9.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_9.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_9.setReadOnly(True)
-        self.itemLabelTextBrowser_9.setObjectName("itemLabelTextBrowser_9")
-        self.horizontalLayout_10.addWidget(self.itemLabelTextBrowser_9)
-        self.SendButton_9 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_9.sizePolicy().hasHeightForWidth())
-        self.SendButton_9.setSizePolicy(sizePolicy)
-        self.SendButton_9.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_9.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_9.setObjectName("SendButton_9")
-        self.horizontalLayout_10.addWidget(self.SendButton_9)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_10)
-        self.horizontalLayout_11 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_11.setObjectName("horizontalLayout_11")
-        self.itemLabelTextBrowser_10 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_10.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_10.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_10.setReadOnly(True)
-        self.itemLabelTextBrowser_10.setObjectName("itemLabelTextBrowser_10")
-        self.horizontalLayout_11.addWidget(self.itemLabelTextBrowser_10)
-        self.SendButton_10 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_10.sizePolicy().hasHeightForWidth())
-        self.SendButton_10.setSizePolicy(sizePolicy)
-        self.SendButton_10.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_10.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_10.setObjectName("SendButton_10")
-        self.horizontalLayout_11.addWidget(self.SendButton_10)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_11)
-        self.horizontalLayout_12 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_12.setObjectName("horizontalLayout_12")
-        self.itemLabelTextBrowser_11 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_11.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_11.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_11.setReadOnly(True)
-        self.itemLabelTextBrowser_11.setObjectName("itemLabelTextBrowser_11")
-        self.horizontalLayout_12.addWidget(self.itemLabelTextBrowser_11)
-        self.SendButton_11 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_11.sizePolicy().hasHeightForWidth())
-        self.SendButton_11.setSizePolicy(sizePolicy)
-        self.SendButton_11.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_11.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_11.setObjectName("SendButton_11")
-        self.horizontalLayout_12.addWidget(self.SendButton_11)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_12)
-        self.horizontalLayout_13 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_13.setObjectName("horizontalLayout_13")
-        self.itemLabelTextBrowser_12 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_12.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_12.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_12.setReadOnly(True)
-        self.itemLabelTextBrowser_12.setObjectName("itemLabelTextBrowser_12")
-        self.horizontalLayout_13.addWidget(self.itemLabelTextBrowser_12)
-        self.SendButton_12 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_12.sizePolicy().hasHeightForWidth())
-        self.SendButton_12.setSizePolicy(sizePolicy)
-        self.SendButton_12.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_12.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_12.setObjectName("SendButton_12")
-        self.horizontalLayout_13.addWidget(self.SendButton_12)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_13)
-        self.horizontalLayout_14 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_14.setObjectName("horizontalLayout_14")
-        self.itemLabelTextBrowser_13 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_13.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_13.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_13.setReadOnly(True)
-        self.itemLabelTextBrowser_13.setObjectName("itemLabelTextBrowser_13")
-        self.horizontalLayout_14.addWidget(self.itemLabelTextBrowser_13)
-        self.SendButton_13 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_13.sizePolicy().hasHeightForWidth())
-        self.SendButton_13.setSizePolicy(sizePolicy)
-        self.SendButton_13.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_13.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_13.setObjectName("SendButton_13")
-        self.horizontalLayout_14.addWidget(self.SendButton_13)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_14)
-        self.horizontalLayout_15 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_15.setObjectName("horizontalLayout_15")
-        self.itemLabelTextBrowser_14 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_14.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_14.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_14.setReadOnly(True)
-        self.itemLabelTextBrowser_14.setObjectName("itemLabelTextBrowser_14")
-        self.horizontalLayout_15.addWidget(self.itemLabelTextBrowser_14)
-        self.SendButton_14 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_14.sizePolicy().hasHeightForWidth())
-        self.SendButton_14.setSizePolicy(sizePolicy)
-        self.SendButton_14.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_14.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_14.setObjectName("SendButton_14")
-        self.horizontalLayout_15.addWidget(self.SendButton_14)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_15)
-        self.horizontalLayout_16 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_16.setObjectName("horizontalLayout_16")
-        self.itemLabelTextBrowser_15 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_15.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_15.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_15.setReadOnly(True)
-        self.itemLabelTextBrowser_15.setObjectName("itemLabelTextBrowser_15")
-        self.horizontalLayout_16.addWidget(self.itemLabelTextBrowser_15)
-        self.SendButton_15 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_15.sizePolicy().hasHeightForWidth())
-        self.SendButton_15.setSizePolicy(sizePolicy)
-        self.SendButton_15.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_15.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_15.setObjectName("SendButton_15")
-        self.horizontalLayout_16.addWidget(self.SendButton_15)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_16)
-        self.horizontalLayout_17 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_17.setObjectName("horizontalLayout_17")
-        self.itemLabelTextBrowser_16 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_16.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_16.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_16.setReadOnly(True)
-        self.itemLabelTextBrowser_16.setObjectName("itemLabelTextBrowser_16")
-        self.horizontalLayout_17.addWidget(self.itemLabelTextBrowser_16)
-        self.SendButton_16 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_16.sizePolicy().hasHeightForWidth())
-        self.SendButton_16.setSizePolicy(sizePolicy)
-        self.SendButton_16.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_16.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_16.setObjectName("SendButton_16")
-        self.horizontalLayout_17.addWidget(self.SendButton_16)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_17)
-        self.horizontalLayout_18 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_18.setObjectName("horizontalLayout_18")
-        self.itemLabelTextBrowser_17 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_17.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_17.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_17.setReadOnly(True)
-        self.itemLabelTextBrowser_17.setObjectName("itemLabelTextBrowser_17")
-        self.horizontalLayout_18.addWidget(self.itemLabelTextBrowser_17)
-        self.SendButton_17 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_17.sizePolicy().hasHeightForWidth())
-        self.SendButton_17.setSizePolicy(sizePolicy)
-        self.SendButton_17.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_17.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_17.setObjectName("SendButton_17")
-        self.horizontalLayout_18.addWidget(self.SendButton_17)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_18)
-        self.horizontalLayout_19 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_19.setObjectName("horizontalLayout_19")
-        self.itemLabelTextBrowser_18 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_18.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_18.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_18.setReadOnly(True)
-        self.itemLabelTextBrowser_18.setObjectName("itemLabelTextBrowser_18")
-        self.horizontalLayout_19.addWidget(self.itemLabelTextBrowser_18)
-        self.SendButton_18 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_18.sizePolicy().hasHeightForWidth())
-        self.SendButton_18.setSizePolicy(sizePolicy)
-        self.SendButton_18.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_18.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_18.setObjectName("SendButton_18")
-        self.horizontalLayout_19.addWidget(self.SendButton_18)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_19)
-        self.horizontalLayout_20 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_20.setObjectName("horizontalLayout_20")
-        self.itemLabelTextBrowser_19 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_19.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_19.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_19.setReadOnly(True)
-        self.itemLabelTextBrowser_19.setObjectName("itemLabelTextBrowser_19")
-        self.horizontalLayout_20.addWidget(self.itemLabelTextBrowser_19)
-        self.SendButton_19 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_19.sizePolicy().hasHeightForWidth())
-        self.SendButton_19.setSizePolicy(sizePolicy)
-        self.SendButton_19.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_19.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_19.setObjectName("SendButton_19")
-        self.horizontalLayout_20.addWidget(self.SendButton_19)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_20)
-        self.horizontalLayout_21 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_21.setObjectName("horizontalLayout_21")
-        self.itemLabelTextBrowser_20 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_20.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_20.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_20.setReadOnly(True)
-        self.itemLabelTextBrowser_20.setObjectName("itemLabelTextBrowser_20")
-        self.horizontalLayout_21.addWidget(self.itemLabelTextBrowser_20)
-        self.SendButton_20 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_20.sizePolicy().hasHeightForWidth())
-        self.SendButton_20.setSizePolicy(sizePolicy)
-        self.SendButton_20.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_20.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_20.setObjectName("SendButton_20")
-        self.horizontalLayout_21.addWidget(self.SendButton_20)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_21)
-        self.horizontalLayout_22 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_22.setObjectName("horizontalLayout_22")
-        self.itemLabelTextBrowser_21 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_21.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_21.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_21.setReadOnly(True)
-        self.itemLabelTextBrowser_21.setObjectName("itemLabelTextBrowser_21")
-        self.horizontalLayout_22.addWidget(self.itemLabelTextBrowser_21)
-        self.SendButton_21 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_21.sizePolicy().hasHeightForWidth())
-        self.SendButton_21.setSizePolicy(sizePolicy)
-        self.SendButton_21.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_21.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_21.setObjectName("SendButton_21")
-        self.horizontalLayout_22.addWidget(self.SendButton_21)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_22)
-        self.horizontalLayout_23 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_23.setObjectName("horizontalLayout_23")
-        self.itemLabelTextBrowser_22 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_22.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_22.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_22.setReadOnly(True)
-        self.itemLabelTextBrowser_22.setObjectName("itemLabelTextBrowser_22")
-        self.horizontalLayout_23.addWidget(self.itemLabelTextBrowser_22)
-        self.SendButton_22 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_22.sizePolicy().hasHeightForWidth())
-        self.SendButton_22.setSizePolicy(sizePolicy)
-        self.SendButton_22.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_22.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_22.setObjectName("SendButton_22")
-        self.horizontalLayout_23.addWidget(self.SendButton_22)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_23)
-        self.horizontalLayout_24 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_24.setObjectName("horizontalLayout_24")
-        self.itemLabelTextBrowser_23 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_23.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_23.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_23.setReadOnly(True)
-        self.itemLabelTextBrowser_23.setObjectName("itemLabelTextBrowser_23")
-        self.horizontalLayout_24.addWidget(self.itemLabelTextBrowser_23)
-        self.SendButton_23 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_23.sizePolicy().hasHeightForWidth())
-        self.SendButton_23.setSizePolicy(sizePolicy)
-        self.SendButton_23.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_23.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_23.setObjectName("SendButton_23")
-        self.horizontalLayout_24.addWidget(self.SendButton_23)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_24)
-        self.horizontalLayout_25 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_25.setObjectName("horizontalLayout_25")
-        self.itemLabelTextBrowser_24 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_24.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_24.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_24.setReadOnly(True)
-        self.itemLabelTextBrowser_24.setObjectName("itemLabelTextBrowser_24")
-        self.horizontalLayout_25.addWidget(self.itemLabelTextBrowser_24)
-        self.SendButton_24 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_24.sizePolicy().hasHeightForWidth())
-        self.SendButton_24.setSizePolicy(sizePolicy)
-        self.SendButton_24.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_24.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_24.setObjectName("SendButton_24")
-        self.horizontalLayout_25.addWidget(self.SendButton_24)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_25)
-        self.horizontalLayout_26 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_26.setObjectName("horizontalLayout_26")
-        self.itemLabelTextBrowser_25 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemLabelTextBrowser_25.setMinimumSize(QtCore.QSize(391, 31))
-        self.itemLabelTextBrowser_25.setMaximumSize(QtCore.QSize(1234, 31))
-        self.itemLabelTextBrowser_25.setReadOnly(True)
-        self.itemLabelTextBrowser_25.setObjectName("itemLabelTextBrowser_25")
-        self.horizontalLayout_26.addWidget(self.itemLabelTextBrowser_25)
-        self.SendButton_25 = QtWidgets.QPushButton(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SendButton_25.sizePolicy().hasHeightForWidth())
-        self.SendButton_25.setSizePolicy(sizePolicy)
-        self.SendButton_25.setMinimumSize(QtCore.QSize(73, 32))
-        self.SendButton_25.setMaximumSize(QtCore.QSize(73, 32))
-        self.SendButton_25.setObjectName("SendButton_25")
-        self.horizontalLayout_26.addWidget(self.SendButton_25)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_26)
-        self.scrollArea.setWidget(self.scrollAreaWidgetContents)
-        self.verticalLayout_6.addWidget(self.scrollArea)
-
-        self.retranslateUi(GenericCommandDialog)
-        self.buttonBox.clicked['QAbstractButton*'].connect(GenericCommandDialog.close)
-        QtCore.QMetaObject.connectSlotsByName(GenericCommandDialog)
-
-    def retranslateUi(self, GenericCommandDialog):
-        _translate = QtCore.QCoreApplication.translate
-        GenericCommandDialog.setWindowTitle(_translate("GenericCommandDialog", "Command Page"))
-        self.subSystemCommandPageLabel.setText(_translate("GenericCommandDialog", "Subsystem"))
-        self.packetIdLabel.setText(_translate("GenericCommandDialog", "Packet ID"))
-        self.label_5.setText(_translate("GenericCommandDialog", "Send To:"))
-        self.label.setText(_translate("GenericCommandDialog", "Command"))
-        self.SendButton_1.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_2.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_3.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_4.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_5.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_6.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_7.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_8.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_9.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_10.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_11.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_12.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_13.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_14.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_15.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_16.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_17.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_18.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_19.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_20.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_21.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_22.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_23.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_24.setText(_translate("GenericCommandDialog", "Send"))
-        self.SendButton_25.setText(_translate("GenericCommandDialog", "Send"))
-
