+++ Subsystems/cmdGui/Ui_GenericCommandDialog.py
@@ -1,0 +1,105 @@
+# -*- coding: utf-8 -*-
+
+# Form implementation generated from reading ui file '/home/lbleier/cFS/tools/cFS-GroundSystem/Subsystems/cmdGui/GenericCommandDialog.ui'
+#
+# Created by: PyQt5 UI code generator 5.10.1
+#
+# WARNING! All changes made in this file will be lost!
+
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+class Ui_GenericCommandDialog(object):
+    def setupUi(self, GenericCommandDialog):
+        GenericCommandDialog.setObjectName("GenericCommandDialog")
+        GenericCommandDialog.resize(549, 1058)
+        self.verticalLayout_6 = QtWidgets.QVBoxLayout(GenericCommandDialog)
+        self.verticalLayout_6.setObjectName("verticalLayout_6")
+        self.verticalLayout_4 = QtWidgets.QVBoxLayout()
+        self.verticalLayout_4.setObjectName("verticalLayout_4")
+        self.horizontalLayout = QtWidgets.QHBoxLayout()
+        self.horizontalLayout.setObjectName("horizontalLayout")
+        self.verticalLayout = QtWidgets.QVBoxLayout()
+        self.verticalLayout.setObjectName("verticalLayout")
+        self.subSystemCommandPageLabel = QtWidgets.QLabel(GenericCommandDialog)
+        self.subSystemCommandPageLabel.setAlignment(QtCore.Qt.AlignCenter)
+        self.subSystemCommandPageLabel.setObjectName("subSystemCommandPageLabel")
+        self.verticalLayout.addWidget(self.subSystemCommandPageLabel)
+        self.subSystemLineEdit = QtWidgets.QLineEdit(GenericCommandDialog)
+        self.subSystemLineEdit.setMinimumSize(QtCore.QSize(159, 31))
+        self.subSystemLineEdit.setMaximumSize(QtCore.QSize(300, 31))
+        self.subSystemLineEdit.setObjectName("subSystemLineEdit")
+        self.verticalLayout.addWidget(self.subSystemLineEdit)
+        self.horizontalLayout.addLayout(self.verticalLayout)
+        self.verticalLayout_2 = QtWidgets.QVBoxLayout()
+        self.verticalLayout_2.setObjectName("verticalLayout_2")
+        self.packetIdLabel = QtWidgets.QLabel(GenericCommandDialog)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Fixed)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.packetIdLabel.sizePolicy().hasHeightForWidth())
+        self.packetIdLabel.setSizePolicy(sizePolicy)
+        self.packetIdLabel.setMinimumSize(QtCore.QSize(0, 13))
+        self.packetIdLabel.setMaximumSize(QtCore.QSize(16777193, 13))
+        self.packetIdLabel.setObjectName("packetIdLabel")
+        self.verticalLayout_2.addWidget(self.packetIdLabel)
+        self.packetId = QtWidgets.QLCDNumber(GenericCommandDialog)
+        self.packetId.setObjectName("packetId")
+        self.verticalLayout_2.addWidget(self.packetId)
+        self.horizontalLayout.addLayout(self.verticalLayout_2)
+        self.verticalLayout_3 = QtWidgets.QVBoxLayout()
+        self.verticalLayout_3.setObjectName("verticalLayout_3")
+        self.label_5 = QtWidgets.QLabel(GenericCommandDialog)
+        self.label_5.setAlignment(QtCore.Qt.AlignCenter)
+        self.label_5.setObjectName("label_5")
+        self.verticalLayout_3.addWidget(self.label_5)
+        self.commandAddressLineEdit = QtWidgets.QLineEdit(GenericCommandDialog)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.commandAddressLineEdit.sizePolicy().hasHeightForWidth())
+        self.commandAddressLineEdit.setSizePolicy(sizePolicy)
+        self.commandAddressLineEdit.setMinimumSize(QtCore.QSize(135, 31))
+        self.commandAddressLineEdit.setMaximumSize(QtCore.QSize(135, 31))
+        self.commandAddressLineEdit.setObjectName("commandAddressLineEdit")
+        self.verticalLayout_3.addWidget(self.commandAddressLineEdit)
+        self.horizontalLayout.addLayout(self.verticalLayout_3)
+        spacerItem = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
+        self.horizontalLayout.addItem(spacerItem)
+        self.buttonBox = QtWidgets.QDialogButtonBox(GenericCommandDialog)
+        self.buttonBox.setOrientation(QtCore.Qt.Horizontal)
+        self.buttonBox.setStandardButtons(QtWidgets.QDialogButtonBox.Close)
+        self.buttonBox.setObjectName("buttonBox")
+        self.horizontalLayout.addWidget(self.buttonBox)
+        self.verticalLayout_4.addLayout(self.horizontalLayout)
+        self.label = QtWidgets.QLabel(GenericCommandDialog)
+        self.label.setAlignment(QtCore.Qt.AlignCenter)
+        self.label.setObjectName("label")
+        self.verticalLayout_4.addWidget(self.label)
+        self.tblCommands = QtWidgets.QTableWidget(GenericCommandDialog)
+        self.tblCommands.setEditTriggers(QtWidgets.QAbstractItemView.NoEditTriggers)
+        self.tblCommands.setObjectName("tblCommands")
+        self.tblCommands.setColumnCount(2)
+        self.tblCommands.setRowCount(0)
+        item = QtWidgets.QTableWidgetItem()
+        self.tblCommands.setHorizontalHeaderItem(0, item)
+        item = QtWidgets.QTableWidgetItem()
+        self.tblCommands.setHorizontalHeaderItem(1, item)
+        self.tblCommands.verticalHeader().setVisible(False)
+        self.tblCommands.verticalHeader().setMinimumSectionSize(40)
+        self.verticalLayout_4.addWidget(self.tblCommands)
+        self.verticalLayout_6.addLayout(self.verticalLayout_4)
+
+        self.retranslateUi(GenericCommandDialog)
+        self.buttonBox.clicked['QAbstractButton*'].connect(GenericCommandDialog.close)
+        QtCore.QMetaObject.connectSlotsByName(GenericCommandDialog)
+
+    def retranslateUi(self, GenericCommandDialog):
+        _translate = QtCore.QCoreApplication.translate
+        GenericCommandDialog.setWindowTitle(_translate("GenericCommandDialog", "Command Page"))
+        self.subSystemCommandPageLabel.setText(_translate("GenericCommandDialog", "Subsystem"))
+        self.packetIdLabel.setText(_translate("GenericCommandDialog", "Packet ID"))
+        self.label_5.setText(_translate("GenericCommandDialog", "Send To:"))
+        self.label.setText(_translate("GenericCommandDialog", "Command"))
+        item = self.tblCommands.horizontalHeaderItem(0)
+        item.setText(_translate("GenericCommandDialog", "Command"))
+
