+++ Subsystems/cmdGui/Ui_ParameterDialog.py
@@ -1,0 +1,94 @@
+# -*- coding: utf-8 -*-
+
+# Form implementation generated from reading ui file '/home/lbleier/cFS/tools/cFS-GroundSystem/Subsystems/cmdGui/ParameterDialog.ui'
+#
+# Created by: PyQt5 UI code generator 5.10.1
+#
+# WARNING! All changes made in this file will be lost!
+
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+class Ui_Dialog(object):
+    def setupUi(self, Dialog):
+        Dialog.setObjectName("Dialog")
+        Dialog.setEnabled(True)
+        Dialog.resize(782, 550)
+        self.label_title = QtWidgets.QLabel(Dialog)
+        self.label_title.setGeometry(QtCore.QRect(330, 120, 91, 31))
+        font = QtGui.QFont()
+        font.setFamily("Sans Serif")
+        font.setPointSize(10)
+        self.label_title.setFont(font)
+        self.label_title.setAlignment(QtCore.Qt.AlignCenter)
+        self.label_title.setObjectName("label_title")
+        self.label_instructions = QtWidgets.QLabel(Dialog)
+        self.label_instructions.setGeometry(QtCore.QRect(120, 140, 551, 31))
+        self.label_instructions.setAlignment(QtCore.Qt.AlignCenter)
+        self.label_instructions.setObjectName("label_instructions")
+        self.buttonBox = QtWidgets.QDialogButtonBox(Dialog)
+        self.buttonBox.setGeometry(QtCore.QRect(670, 490, 101, 31))
+        self.buttonBox.setLayoutDirection(QtCore.Qt.LeftToRight)
+        self.buttonBox.setOrientation(QtCore.Qt.Horizontal)
+        self.buttonBox.setStandardButtons(QtWidgets.QDialogButtonBox.Close)
+        self.buttonBox.setCenterButtons(True)
+        self.buttonBox.setObjectName("buttonBox")
+        self.status_box = QtWidgets.QTextBrowser(Dialog)
+        self.status_box.setGeometry(QtCore.QRect(480, 40, 201, 41))
+        self.status_box.setAutoFillBackground(False)
+        self.status_box.setObjectName("status_box")
+        self.label_param_title_2 = QtWidgets.QLabel(Dialog)
+        self.label_param_title_2.setGeometry(QtCore.QRect(480, 10, 61, 21))
+        self.label_param_title_2.setObjectName("label_param_title_2")
+        self.SendButton_1 = QtWidgets.QPushButton(Dialog)
+        self.SendButton_1.setGeometry(QtCore.QRect(690, 47, 71, 27))
+        self.SendButton_1.setAutoDefault(False)
+        self.SendButton_1.setDefault(True)
+        self.SendButton_1.setObjectName("SendButton_1")
+        self.label_5 = QtWidgets.QLabel(Dialog)
+        self.label_5.setGeometry(QtCore.QRect(260, 10, 81, 20))
+        self.label_5.setObjectName("label_5")
+        self.subSystemCommandPageLabel = QtWidgets.QLabel(Dialog)
+        self.subSystemCommandPageLabel.setGeometry(QtCore.QRect(30, 10, 91, 24))
+        self.subSystemCommandPageLabel.setObjectName("subSystemCommandPageLabel")
+        self.subSystemTextBrowser = QtWidgets.QTextBrowser(Dialog)
+        self.subSystemTextBrowser.setGeometry(QtCore.QRect(30, 40, 221, 41))
+        self.subSystemTextBrowser.setObjectName("subSystemTextBrowser")
+        self.commandAddressTextBrowser = QtWidgets.QTextBrowser(Dialog)
+        self.commandAddressTextBrowser.setGeometry(QtCore.QRect(260, 40, 211, 41))
+        self.commandAddressTextBrowser.setObjectName("commandAddressTextBrowser")
+        self.tblParameters = QtWidgets.QTableWidget(Dialog)
+        self.tblParameters.setGeometry(QtCore.QRect(20, 180, 731, 301))
+        self.tblParameters.setObjectName("tblParameters")
+        self.tblParameters.setColumnCount(3)
+        self.tblParameters.setRowCount(0)
+        item = QtWidgets.QTableWidgetItem()
+        self.tblParameters.setHorizontalHeaderItem(0, item)
+        item = QtWidgets.QTableWidgetItem()
+        self.tblParameters.setHorizontalHeaderItem(1, item)
+        item = QtWidgets.QTableWidgetItem()
+        self.tblParameters.setHorizontalHeaderItem(2, item)
+        self.tblParameters.verticalHeader().setVisible(False)
+
+        self.retranslateUi(Dialog)
+        self.buttonBox.clicked['QAbstractButton*'].connect(Dialog.close)
+        QtCore.QMetaObject.connectSlotsByName(Dialog)
+        Dialog.setTabOrder(self.status_box, self.SendButton_1)
+        Dialog.setTabOrder(self.SendButton_1, self.subSystemTextBrowser)
+        Dialog.setTabOrder(self.subSystemTextBrowser, self.commandAddressTextBrowser)
+
+    def retranslateUi(self, Dialog):
+        _translate = QtCore.QCoreApplication.translate
+        Dialog.setWindowTitle(_translate("Dialog", "Dialog"))
+        self.label_title.setText(_translate("Dialog", "Parameters"))
+        self.label_instructions.setText(_translate("Dialog", "Please enter the following parameters then click \'Send\':"))
+        self.label_param_title_2.setText(_translate("Dialog", "Status:"))
+        self.SendButton_1.setText(_translate("Dialog", "Send"))
+        self.label_5.setText(_translate("Dialog", "Command:"))
+        self.subSystemCommandPageLabel.setText(_translate("Dialog", "Subsystem:"))
+        item = self.tblParameters.horizontalHeaderItem(0)
+        item.setText(_translate("Dialog", "Parameter"))
+        item = self.tblParameters.horizontalHeaderItem(1)
+        item.setText(_translate("Dialog", "Description"))
+        item = self.tblParameters.horizontalHeaderItem(2)
+        item.setText(_translate("Dialog", "Input"))
+
