--- Subsystems/tlmGUI/GenericTelemetryDialog.py
@@ -1,861 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file '/home/lbleier/cFS/tools/cFS-GroundSystem/Subsystems/tlmGUI/GenericTelemetryDialog.ui'
-#
-# Created by: PyQt5 UI code generator 5.10.1
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-class Ui_GenericTelemetryDialog(object):
-    def setupUi(self, GenericTelemetryDialog):
-        GenericTelemetryDialog.setObjectName("GenericTelemetryDialog")
-        GenericTelemetryDialog.resize(1046, 816)
-        self.layoutWidget = QtWidgets.QWidget(GenericTelemetryDialog)
-        self.layoutWidget.setGeometry(QtCore.QRect(40, 22, 841, 65))
-        self.layoutWidget.setObjectName("layoutWidget")
-        self.horizontalLayout = QtWidgets.QHBoxLayout(self.layoutWidget)
-        self.horizontalLayout.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout.setObjectName("horizontalLayout")
-        self.verticalLayout_2 = QtWidgets.QVBoxLayout()
-        self.verticalLayout_2.setObjectName("verticalLayout_2")
-        self.subSystemTelemetryPageLabel = QtWidgets.QLabel(self.layoutWidget)
-        self.subSystemTelemetryPageLabel.setObjectName("subSystemTelemetryPageLabel")
-        self.verticalLayout_2.addWidget(self.subSystemTelemetryPageLabel)
-        self.subSystemLineEdit = QtWidgets.QLineEdit(self.layoutWidget)
-        self.subSystemLineEdit.setMinimumSize(QtCore.QSize(141, 31))
-        self.subSystemLineEdit.setMaximumSize(QtCore.QSize(141, 31))
-        self.subSystemLineEdit.setReadOnly(True)
-        self.subSystemLineEdit.setObjectName("subSystemLineEdit")
-        self.verticalLayout_2.addWidget(self.subSystemLineEdit)
-        self.horizontalLayout.addLayout(self.verticalLayout_2)
-        spacerItem = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem)
-        self.verticalLayout = QtWidgets.QVBoxLayout()
-        self.verticalLayout.setObjectName("verticalLayout")
-        self.packetIdLabel = QtWidgets.QLabel(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.packetIdLabel.sizePolicy().hasHeightForWidth())
-        self.packetIdLabel.setSizePolicy(sizePolicy)
-        self.packetIdLabel.setMinimumSize(QtCore.QSize(0, 21))
-        self.packetIdLabel.setMaximumSize(QtCore.QSize(16777215, 25))
-        self.packetIdLabel.setObjectName("packetIdLabel")
-        self.verticalLayout.addWidget(self.packetIdLabel)
-        self.packetId = QtWidgets.QLCDNumber(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Minimum)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.packetId.sizePolicy().hasHeightForWidth())
-        self.packetId.setSizePolicy(sizePolicy)
-        self.packetId.setObjectName("packetId")
-        self.verticalLayout.addWidget(self.packetId)
-        self.horizontalLayout.addLayout(self.verticalLayout)
-        spacerItem1 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem1)
-        self.verticalLayout_3 = QtWidgets.QVBoxLayout()
-        self.verticalLayout_3.setObjectName("verticalLayout_3")
-        self.label_5 = QtWidgets.QLabel(self.layoutWidget)
-        self.label_5.setObjectName("label_5")
-        self.verticalLayout_3.addWidget(self.label_5)
-        self.sequenceCount = QtWidgets.QSpinBox(self.layoutWidget)
-        self.sequenceCount.setMinimumSize(QtCore.QSize(101, 31))
-        self.sequenceCount.setMaximumSize(QtCore.QSize(101, 31))
-        self.sequenceCount.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.sequenceCount.setMaximum(16384)
-        self.sequenceCount.setProperty("value", 0)
-        self.sequenceCount.setObjectName("sequenceCount")
-        self.verticalLayout_3.addWidget(self.sequenceCount)
-        self.horizontalLayout.addLayout(self.verticalLayout_3)
-        spacerItem2 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem2)
-        self.buttonBox = QtWidgets.QDialogButtonBox(self.layoutWidget)
-        self.buttonBox.setStandardButtons(QtWidgets.QDialogButtonBox.Close)
-        self.buttonBox.setObjectName("buttonBox")
-        self.horizontalLayout.addWidget(self.buttonBox)
-        self.label_6 = QtWidgets.QLabel(GenericTelemetryDialog)
-        self.label_6.setGeometry(QtCore.QRect(29, 750, 711, 19))
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.label_6.setFont(font)
-        self.label_6.setObjectName("label_6")
-        self.layoutWidget_2 = QtWidgets.QWidget(GenericTelemetryDialog)
-        self.layoutWidget_2.setGeometry(QtCore.QRect(40, 100, 451, 23))
-        self.layoutWidget_2.setObjectName("layoutWidget_2")
-        self.horizontalLayout_4 = QtWidgets.QHBoxLayout(self.layoutWidget_2)
-        self.horizontalLayout_4.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_4.setObjectName("horizontalLayout_4")
-        self.label = QtWidgets.QLabel(self.layoutWidget_2)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label.sizePolicy().hasHeightForWidth())
-        self.label.setSizePolicy(sizePolicy)
-        self.label.setMinimumSize(QtCore.QSize(161, 17))
-        self.label.setAlignment(QtCore.Qt.AlignCenter)
-        self.label.setObjectName("label")
-        self.horizontalLayout_4.addWidget(self.label)
-        self.label_8 = QtWidgets.QLabel(self.layoutWidget_2)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_8.sizePolicy().hasHeightForWidth())
-        self.label_8.setSizePolicy(sizePolicy)
-        self.label_8.setMinimumSize(QtCore.QSize(161, 17))
-        self.label_8.setAlignment(QtCore.Qt.AlignCenter)
-        self.label_8.setObjectName("label_8")
-        self.horizontalLayout_4.addWidget(self.label_8)
-        self.scrollArea = QtWidgets.QScrollArea(GenericTelemetryDialog)
-        self.scrollArea.setGeometry(QtCore.QRect(30, 130, 961, 601))
-        self.scrollArea.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOn)
-        self.scrollArea.setWidgetResizable(True)
-        self.scrollArea.setObjectName("scrollArea")
-        self.scrollAreaWidgetContents = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents.setGeometry(QtCore.QRect(0, 0, 945, 1000))
-        self.scrollAreaWidgetContents.setMinimumSize(QtCore.QSize(0, 1000))
-        self.scrollAreaWidgetContents.setObjectName("scrollAreaWidgetContents")
-        self.horizontalLayout_3 = QtWidgets.QHBoxLayout(self.scrollAreaWidgetContents)
-        self.horizontalLayout_3.setObjectName("horizontalLayout_3")
-        self.verticalLayout_4 = QtWidgets.QVBoxLayout()
-        self.verticalLayout_4.setObjectName("verticalLayout_4")
-        self.itemLabelPlainTextEdit_1 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_1.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_1.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_1.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_1.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_1.setReadOnly(True)
-        self.itemLabelPlainTextEdit_1.setObjectName("itemLabelPlainTextEdit_1")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_1)
-        self.itemLabelPlainTextEdit_2 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_2.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_2.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_2.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_2.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_2.setReadOnly(True)
-        self.itemLabelPlainTextEdit_2.setObjectName("itemLabelPlainTextEdit_2")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_2)
-        self.itemLabelPlainTextEdit_3 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_3.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_3.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_3.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_3.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_3.setReadOnly(True)
-        self.itemLabelPlainTextEdit_3.setObjectName("itemLabelPlainTextEdit_3")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_3)
-        self.itemLabelPlainTextEdit_4 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_4.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_4.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_4.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_4.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_4.setReadOnly(True)
-        self.itemLabelPlainTextEdit_4.setObjectName("itemLabelPlainTextEdit_4")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_4)
-        self.itemLabelPlainTextEdit_5 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_5.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_5.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_5.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_5.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_5.setReadOnly(True)
-        self.itemLabelPlainTextEdit_5.setObjectName("itemLabelPlainTextEdit_5")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_5)
-        self.itemLabelPlainTextEdit_6 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_6.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_6.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_6.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_6.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_6.setReadOnly(True)
-        self.itemLabelPlainTextEdit_6.setObjectName("itemLabelPlainTextEdit_6")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_6)
-        self.itemLabelPlainTextEdit_7 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_7.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_7.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_7.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_7.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_7.setReadOnly(True)
-        self.itemLabelPlainTextEdit_7.setObjectName("itemLabelPlainTextEdit_7")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_7)
-        self.itemLabelPlainTextEdit_8 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_8.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_8.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_8.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_8.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_8.setReadOnly(True)
-        self.itemLabelPlainTextEdit_8.setObjectName("itemLabelPlainTextEdit_8")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_8)
-        self.itemLabelPlainTextEdit_9 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_9.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_9.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_9.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_9.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_9.setReadOnly(True)
-        self.itemLabelPlainTextEdit_9.setObjectName("itemLabelPlainTextEdit_9")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_9)
-        self.itemLabelPlainTextEdit_10 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_10.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_10.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_10.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_10.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_10.setReadOnly(True)
-        self.itemLabelPlainTextEdit_10.setObjectName("itemLabelPlainTextEdit_10")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_10)
-        self.itemLabelPlainTextEdit_11 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_11.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_11.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_11.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_11.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_11.setReadOnly(True)
-        self.itemLabelPlainTextEdit_11.setObjectName("itemLabelPlainTextEdit_11")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_11)
-        self.itemLabelPlainTextEdit_12 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_12.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_12.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_12.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_12.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_12.setReadOnly(True)
-        self.itemLabelPlainTextEdit_12.setObjectName("itemLabelPlainTextEdit_12")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_12)
-        self.itemLabelPlainTextEdit_13 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_13.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_13.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_13.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_13.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_13.setReadOnly(True)
-        self.itemLabelPlainTextEdit_13.setObjectName("itemLabelPlainTextEdit_13")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_13)
-        self.itemLabelPlainTextEdit_14 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_14.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_14.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_14.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_14.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_14.setReadOnly(True)
-        self.itemLabelPlainTextEdit_14.setObjectName("itemLabelPlainTextEdit_14")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_14)
-        self.itemLabelPlainTextEdit_15 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_15.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_15.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_15.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_15.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_15.setReadOnly(True)
-        self.itemLabelPlainTextEdit_15.setObjectName("itemLabelPlainTextEdit_15")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_15)
-        self.itemLabelPlainTextEdit_16 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_16.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_16.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_16.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_16.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_16.setReadOnly(True)
-        self.itemLabelPlainTextEdit_16.setObjectName("itemLabelPlainTextEdit_16")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_16)
-        self.itemLabelPlainTextEdit_17 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_17.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_17.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_17.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_17.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_17.setReadOnly(True)
-        self.itemLabelPlainTextEdit_17.setObjectName("itemLabelPlainTextEdit_17")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_17)
-        self.itemLabelPlainTextEdit_18 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_18.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_18.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_18.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_18.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_18.setReadOnly(True)
-        self.itemLabelPlainTextEdit_18.setObjectName("itemLabelPlainTextEdit_18")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_18)
-        self.itemLabelPlainTextEdit_19 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_19.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_19.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_19.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_19.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_19.setReadOnly(True)
-        self.itemLabelPlainTextEdit_19.setObjectName("itemLabelPlainTextEdit_19")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_19)
-        self.itemLabelPlainTextEdit_20 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_20.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_20.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_20.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_20.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_20.setReadOnly(True)
-        self.itemLabelPlainTextEdit_20.setObjectName("itemLabelPlainTextEdit_20")
-        self.verticalLayout_4.addWidget(self.itemLabelPlainTextEdit_20)
-        self.horizontalLayout_3.addLayout(self.verticalLayout_4)
-        self.verticalLayout_5 = QtWidgets.QVBoxLayout()
-        self.verticalLayout_5.setObjectName("verticalLayout_5")
-        self.itemValueLineEdit_1 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_1.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_1.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_1.setReadOnly(True)
-        self.itemValueLineEdit_1.setObjectName("itemValueLineEdit_1")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_1)
-        self.itemValueLineEdit_2 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_2.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_2.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_2.setReadOnly(True)
-        self.itemValueLineEdit_2.setObjectName("itemValueLineEdit_2")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_2)
-        self.itemValueLineEdit_3 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_3.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_3.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_3.setReadOnly(True)
-        self.itemValueLineEdit_3.setObjectName("itemValueLineEdit_3")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_3)
-        self.itemValueLineEdit_4 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_4.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_4.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_4.setReadOnly(True)
-        self.itemValueLineEdit_4.setObjectName("itemValueLineEdit_4")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_4)
-        self.itemValueLineEdit_5 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_5.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_5.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_5.setReadOnly(True)
-        self.itemValueLineEdit_5.setObjectName("itemValueLineEdit_5")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_5)
-        self.itemValueLineEdit_6 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_6.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_6.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_6.setReadOnly(True)
-        self.itemValueLineEdit_6.setObjectName("itemValueLineEdit_6")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_6)
-        self.itemValueLineEdit_7 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_7.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_7.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_7.setReadOnly(True)
-        self.itemValueLineEdit_7.setObjectName("itemValueLineEdit_7")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_7)
-        self.itemValueLineEdit_8 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_8.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_8.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_8.setReadOnly(True)
-        self.itemValueLineEdit_8.setObjectName("itemValueLineEdit_8")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_8)
-        self.itemValueLineEdit_9 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_9.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_9.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_9.setReadOnly(True)
-        self.itemValueLineEdit_9.setObjectName("itemValueLineEdit_9")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_9)
-        self.itemValueLineEdit_10 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_10.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_10.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_10.setReadOnly(True)
-        self.itemValueLineEdit_10.setObjectName("itemValueLineEdit_10")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_10)
-        self.itemValueLineEdit_11 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_11.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_11.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_11.setReadOnly(True)
-        self.itemValueLineEdit_11.setObjectName("itemValueLineEdit_11")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_11)
-        self.itemValueLineEdit_12 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_12.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_12.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_12.setReadOnly(True)
-        self.itemValueLineEdit_12.setObjectName("itemValueLineEdit_12")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_12)
-        self.itemValueLineEdit_13 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_13.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_13.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_13.setReadOnly(True)
-        self.itemValueLineEdit_13.setObjectName("itemValueLineEdit_13")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_13)
-        self.itemValueLineEdit_14 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_14.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_14.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_14.setReadOnly(True)
-        self.itemValueLineEdit_14.setObjectName("itemValueLineEdit_14")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_14)
-        self.itemValueLineEdit_15 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_15.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_15.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_15.setReadOnly(True)
-        self.itemValueLineEdit_15.setObjectName("itemValueLineEdit_15")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_15)
-        self.itemValueLineEdit_16 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_16.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_16.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_16.setReadOnly(True)
-        self.itemValueLineEdit_16.setObjectName("itemValueLineEdit_16")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_16)
-        self.itemValueLineEdit_17 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_17.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_17.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_17.setReadOnly(True)
-        self.itemValueLineEdit_17.setObjectName("itemValueLineEdit_17")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_17)
-        self.itemValueLineEdit_18 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_18.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_18.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_18.setReadOnly(True)
-        self.itemValueLineEdit_18.setObjectName("itemValueLineEdit_18")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_18)
-        self.itemValueLineEdit_19 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_19.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_19.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_19.setReadOnly(True)
-        self.itemValueLineEdit_19.setObjectName("itemValueLineEdit_19")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_19)
-        self.itemValueLineEdit_20 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_20.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_20.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_20.setReadOnly(True)
-        self.itemValueLineEdit_20.setObjectName("itemValueLineEdit_20")
-        self.verticalLayout_5.addWidget(self.itemValueLineEdit_20)
-        self.horizontalLayout_3.addLayout(self.verticalLayout_5)
-        self.line = QtWidgets.QFrame(self.scrollAreaWidgetContents)
-        self.line.setFrameShape(QtWidgets.QFrame.VLine)
-        self.line.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line.setObjectName("line")
-        self.horizontalLayout_3.addWidget(self.line)
-        self.verticalLayout_6 = QtWidgets.QVBoxLayout()
-        self.verticalLayout_6.setObjectName("verticalLayout_6")
-        self.itemLabelPlainTextEdit_21 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_21.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_21.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_21.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_21.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_21.setReadOnly(True)
-        self.itemLabelPlainTextEdit_21.setObjectName("itemLabelPlainTextEdit_21")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_21)
-        self.itemLabelPlainTextEdit_22 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_22.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_22.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_22.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_22.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_22.setReadOnly(True)
-        self.itemLabelPlainTextEdit_22.setObjectName("itemLabelPlainTextEdit_22")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_22)
-        self.itemLabelPlainTextEdit_23 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_23.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_23.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_23.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_23.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_23.setReadOnly(True)
-        self.itemLabelPlainTextEdit_23.setObjectName("itemLabelPlainTextEdit_23")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_23)
-        self.itemLabelPlainTextEdit_24 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_24.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_24.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_24.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_24.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_24.setReadOnly(True)
-        self.itemLabelPlainTextEdit_24.setObjectName("itemLabelPlainTextEdit_24")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_24)
-        self.itemLabelPlainTextEdit_25 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_25.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_25.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_25.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_25.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_25.setReadOnly(True)
-        self.itemLabelPlainTextEdit_25.setObjectName("itemLabelPlainTextEdit_25")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_25)
-        self.itemLabelPlainTextEdit_26 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_26.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_26.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_26.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_26.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_26.setReadOnly(True)
-        self.itemLabelPlainTextEdit_26.setObjectName("itemLabelPlainTextEdit_26")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_26)
-        self.itemLabelPlainTextEdit_27 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_27.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_27.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_27.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_27.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_27.setReadOnly(True)
-        self.itemLabelPlainTextEdit_27.setObjectName("itemLabelPlainTextEdit_27")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_27)
-        self.itemLabelPlainTextEdit_28 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_28.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_28.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_28.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_28.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_28.setReadOnly(True)
-        self.itemLabelPlainTextEdit_28.setObjectName("itemLabelPlainTextEdit_28")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_28)
-        self.itemLabelPlainTextEdit_29 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_29.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_29.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_29.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_29.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_29.setReadOnly(True)
-        self.itemLabelPlainTextEdit_29.setObjectName("itemLabelPlainTextEdit_29")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_29)
-        self.itemLabelPlainTextEdit_30 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_30.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_30.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_30.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_30.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_30.setReadOnly(True)
-        self.itemLabelPlainTextEdit_30.setObjectName("itemLabelPlainTextEdit_30")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_30)
-        self.itemLabelPlainTextEdit_31 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_31.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_31.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_31.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_31.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_31.setReadOnly(True)
-        self.itemLabelPlainTextEdit_31.setObjectName("itemLabelPlainTextEdit_31")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_31)
-        self.itemLabelPlainTextEdit_32 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_32.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_32.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_32.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_32.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_32.setReadOnly(True)
-        self.itemLabelPlainTextEdit_32.setObjectName("itemLabelPlainTextEdit_32")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_32)
-        self.itemLabelPlainTextEdit_33 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_33.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_33.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_33.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_33.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_33.setReadOnly(True)
-        self.itemLabelPlainTextEdit_33.setObjectName("itemLabelPlainTextEdit_33")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_33)
-        self.itemLabelPlainTextEdit_34 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_34.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_34.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_34.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_34.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_34.setReadOnly(True)
-        self.itemLabelPlainTextEdit_34.setObjectName("itemLabelPlainTextEdit_34")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_34)
-        self.itemLabelPlainTextEdit_35 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_35.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_35.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_35.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_35.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_35.setReadOnly(True)
-        self.itemLabelPlainTextEdit_35.setObjectName("itemLabelPlainTextEdit_35")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_35)
-        self.itemLabelPlainTextEdit_36 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_36.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_36.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_36.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_36.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_36.setReadOnly(True)
-        self.itemLabelPlainTextEdit_36.setObjectName("itemLabelPlainTextEdit_36")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_36)
-        self.itemLabelPlainTextEdit_37 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_37.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_37.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_37.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_37.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_37.setReadOnly(True)
-        self.itemLabelPlainTextEdit_37.setObjectName("itemLabelPlainTextEdit_37")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_37)
-        self.itemLabelPlainTextEdit_38 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_38.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_38.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_38.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_38.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_38.setReadOnly(True)
-        self.itemLabelPlainTextEdit_38.setObjectName("itemLabelPlainTextEdit_38")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_38)
-        self.itemLabelPlainTextEdit_39 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_39.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_39.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_39.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_39.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_39.setReadOnly(True)
-        self.itemLabelPlainTextEdit_39.setObjectName("itemLabelPlainTextEdit_39")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_39)
-        self.itemLabelPlainTextEdit_40 = QtWidgets.QPlainTextEdit(self.scrollAreaWidgetContents)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.itemLabelPlainTextEdit_40.sizePolicy().hasHeightForWidth())
-        self.itemLabelPlainTextEdit_40.setSizePolicy(sizePolicy)
-        self.itemLabelPlainTextEdit_40.setMinimumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_40.setMaximumSize(QtCore.QSize(161, 31))
-        self.itemLabelPlainTextEdit_40.setReadOnly(True)
-        self.itemLabelPlainTextEdit_40.setObjectName("itemLabelPlainTextEdit_40")
-        self.verticalLayout_6.addWidget(self.itemLabelPlainTextEdit_40)
-        self.horizontalLayout_3.addLayout(self.verticalLayout_6)
-        self.verticalLayout_7 = QtWidgets.QVBoxLayout()
-        self.verticalLayout_7.setObjectName("verticalLayout_7")
-        self.itemValueLineEdit_21 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_21.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_21.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_21.setReadOnly(True)
-        self.itemValueLineEdit_21.setObjectName("itemValueLineEdit_21")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_21)
-        self.itemValueLineEdit_22 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_22.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_22.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_22.setReadOnly(True)
-        self.itemValueLineEdit_22.setObjectName("itemValueLineEdit_22")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_22)
-        self.itemValueLineEdit_23 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_23.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_23.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_23.setReadOnly(True)
-        self.itemValueLineEdit_23.setObjectName("itemValueLineEdit_23")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_23)
-        self.itemValueLineEdit_24 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_24.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_24.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_24.setReadOnly(True)
-        self.itemValueLineEdit_24.setObjectName("itemValueLineEdit_24")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_24)
-        self.itemValueLineEdit_25 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_25.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_25.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_25.setReadOnly(True)
-        self.itemValueLineEdit_25.setObjectName("itemValueLineEdit_25")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_25)
-        self.itemValueLineEdit_26 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_26.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_26.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_26.setReadOnly(True)
-        self.itemValueLineEdit_26.setObjectName("itemValueLineEdit_26")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_26)
-        self.itemValueLineEdit_27 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_27.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_27.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_27.setReadOnly(True)
-        self.itemValueLineEdit_27.setObjectName("itemValueLineEdit_27")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_27)
-        self.itemValueLineEdit_28 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_28.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_28.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_28.setReadOnly(True)
-        self.itemValueLineEdit_28.setObjectName("itemValueLineEdit_28")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_28)
-        self.itemValueLineEdit_29 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_29.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_29.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_29.setReadOnly(True)
-        self.itemValueLineEdit_29.setObjectName("itemValueLineEdit_29")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_29)
-        self.itemValueLineEdit_30 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_30.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_30.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_30.setReadOnly(True)
-        self.itemValueLineEdit_30.setObjectName("itemValueLineEdit_30")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_30)
-        self.itemValueLineEdit_31 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_31.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_31.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_31.setReadOnly(True)
-        self.itemValueLineEdit_31.setObjectName("itemValueLineEdit_31")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_31)
-        self.itemValueLineEdit_32 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_32.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_32.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_32.setReadOnly(True)
-        self.itemValueLineEdit_32.setObjectName("itemValueLineEdit_32")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_32)
-        self.itemValueLineEdit_33 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_33.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_33.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_33.setReadOnly(True)
-        self.itemValueLineEdit_33.setObjectName("itemValueLineEdit_33")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_33)
-        self.itemValueLineEdit_34 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_34.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_34.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_34.setReadOnly(True)
-        self.itemValueLineEdit_34.setObjectName("itemValueLineEdit_34")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_34)
-        self.itemValueLineEdit_35 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_35.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_35.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_35.setReadOnly(True)
-        self.itemValueLineEdit_35.setObjectName("itemValueLineEdit_35")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_35)
-        self.itemValueLineEdit_36 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_36.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_36.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_36.setReadOnly(True)
-        self.itemValueLineEdit_36.setObjectName("itemValueLineEdit_36")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_36)
-        self.itemValueLineEdit_37 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_37.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_37.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_37.setReadOnly(True)
-        self.itemValueLineEdit_37.setObjectName("itemValueLineEdit_37")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_37)
-        self.itemValueLineEdit_38 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_38.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_38.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_38.setReadOnly(True)
-        self.itemValueLineEdit_38.setObjectName("itemValueLineEdit_38")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_38)
-        self.itemValueLineEdit_39 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_39.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_39.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_39.setReadOnly(True)
-        self.itemValueLineEdit_39.setObjectName("itemValueLineEdit_39")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_39)
-        self.itemValueLineEdit_40 = QtWidgets.QLineEdit(self.scrollAreaWidgetContents)
-        self.itemValueLineEdit_40.setMinimumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_40.setMaximumSize(QtCore.QSize(151, 31))
-        self.itemValueLineEdit_40.setReadOnly(True)
-        self.itemValueLineEdit_40.setObjectName("itemValueLineEdit_40")
-        self.verticalLayout_7.addWidget(self.itemValueLineEdit_40)
-        self.horizontalLayout_3.addLayout(self.verticalLayout_7)
-        self.scrollArea.setWidget(self.scrollAreaWidgetContents)
-        self.layoutWidget_3 = QtWidgets.QWidget(GenericTelemetryDialog)
-        self.layoutWidget_3.setGeometry(QtCore.QRect(510, 100, 451, 23))
-        self.layoutWidget_3.setObjectName("layoutWidget_3")
-        self.horizontalLayout_5 = QtWidgets.QHBoxLayout(self.layoutWidget_3)
-        self.horizontalLayout_5.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_5.setObjectName("horizontalLayout_5")
-        self.label_2 = QtWidgets.QLabel(self.layoutWidget_3)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_2.sizePolicy().hasHeightForWidth())
-        self.label_2.setSizePolicy(sizePolicy)
-        self.label_2.setMinimumSize(QtCore.QSize(161, 17))
-        self.label_2.setAlignment(QtCore.Qt.AlignCenter)
-        self.label_2.setObjectName("label_2")
-        self.horizontalLayout_5.addWidget(self.label_2)
-        self.label_9 = QtWidgets.QLabel(self.layoutWidget_3)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_9.sizePolicy().hasHeightForWidth())
-        self.label_9.setSizePolicy(sizePolicy)
-        self.label_9.setMinimumSize(QtCore.QSize(161, 17))
-        self.label_9.setAlignment(QtCore.Qt.AlignCenter)
-        self.label_9.setObjectName("label_9")
-        self.horizontalLayout_5.addWidget(self.label_9)
-
-        self.retranslateUi(GenericTelemetryDialog)
-        self.buttonBox.clicked['QAbstractButton*'].connect(GenericTelemetryDialog.close)
-        QtCore.QMetaObject.connectSlotsByName(GenericTelemetryDialog)
-
-    def retranslateUi(self, GenericTelemetryDialog):
-        _translate = QtCore.QCoreApplication.translate
-        GenericTelemetryDialog.setWindowTitle(_translate("GenericTelemetryDialog", "Telemetry Page"))
-        self.subSystemTelemetryPageLabel.setText(_translate("GenericTelemetryDialog", "Subsystem Telemetry Page"))
-        self.packetIdLabel.setText(_translate("GenericTelemetryDialog", "Packet ID"))
-        self.label_5.setText(_translate("GenericTelemetryDialog", "Sequence Count"))
-        self.label_6.setText(_translate("GenericTelemetryDialog", "*No packets? Remember to select the IP address of your spacecraft in the Main Window."))
-        self.label.setText(_translate("GenericTelemetryDialog", "Telemetry Point Label"))
-        self.label_8.setText(_translate("GenericTelemetryDialog", "Telemetry Point Value"))
-        self.label_2.setText(_translate("GenericTelemetryDialog", "Telemetry Point Label"))
-        self.label_9.setText(_translate("GenericTelemetryDialog", "Telemetry Point Value"))
-
