--- Subsystems/tlmGUI/TelemetrySystemDialog.py
@@ -1,912 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file '/home/lbleier/cFS/tools/cFS-GroundSystem/Subsystems/tlmGUI/TelemetrySystemDialog.ui'
-#
-# Created by: PyQt5 UI code generator 5.10.1
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-class Ui_TelemetrySystemDialog(object):
-    def setupUi(self, TelemetrySystemDialog):
-        TelemetrySystemDialog.setObjectName("TelemetrySystemDialog")
-        TelemetrySystemDialog.resize(625, 908)
-        self.verticalLayout = QtWidgets.QVBoxLayout(TelemetrySystemDialog)
-        self.verticalLayout.setObjectName("verticalLayout")
-        self.horizontalLayout = QtWidgets.QHBoxLayout()
-        self.horizontalLayout.setObjectName("horizontalLayout")
-        spacerItem = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem)
-        self.label = QtWidgets.QLabel(TelemetrySystemDialog)
-        self.label.setObjectName("label")
-        self.horizontalLayout.addWidget(self.label)
-        spacerItem1 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem1)
-        self.verticalLayout.addLayout(self.horizontalLayout)
-        self.horizontalLayout_3 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_3.setSizeConstraint(QtWidgets.QLayout.SetNoConstraint)
-        self.horizontalLayout_3.setObjectName("horizontalLayout_3")
-        spacerItem2 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_3.addItem(spacerItem2)
-        spacerItem3 = QtWidgets.QSpacerItem(80, 32, QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_3.addItem(spacerItem3)
-        self.horizontalLayout_2 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_2.setObjectName("horizontalLayout_2")
-        spacerItem4 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_2.addItem(spacerItem4)
-        self.label_6 = QtWidgets.QLabel(TelemetrySystemDialog)
-        self.label_6.setObjectName("label_6")
-        self.horizontalLayout_2.addWidget(self.label_6)
-        self.packetCount = QtWidgets.QSpinBox(TelemetrySystemDialog)
-        self.packetCount.setReadOnly(True)
-        self.packetCount.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.packetCount.setMaximum(16384)
-        self.packetCount.setProperty("value", 0)
-        self.packetCount.setObjectName("packetCount")
-        self.horizontalLayout_2.addWidget(self.packetCount)
-        spacerItem5 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_2.addItem(spacerItem5)
-        self.horizontalLayout_3.addLayout(self.horizontalLayout_2)
-        spacerItem6 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_3.addItem(spacerItem6)
-        self.buttonBox = QtWidgets.QDialogButtonBox(TelemetrySystemDialog)
-        self.buttonBox.setStandardButtons(QtWidgets.QDialogButtonBox.Close)
-        self.buttonBox.setObjectName("buttonBox")
-        self.horizontalLayout_3.addWidget(self.buttonBox)
-        self.verticalLayout.addLayout(self.horizontalLayout_3)
-        self.horizontalLayout_4 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_4.setObjectName("horizontalLayout_4")
-        spacerItem7 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_4.addItem(spacerItem7)
-        self.label_2 = QtWidgets.QLabel(TelemetrySystemDialog)
-        self.label_2.setObjectName("label_2")
-        self.horizontalLayout_4.addWidget(self.label_2)
-        spacerItem8 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_4.addItem(spacerItem8)
-        self.verticalLayout.addLayout(self.horizontalLayout_4)
-        self.scrollArea = QtWidgets.QScrollArea(TelemetrySystemDialog)
-        self.scrollArea.setMinimumSize(QtCore.QSize(0, 0))
-        self.scrollArea.setBaseSize(QtCore.QSize(0, 800))
-        self.scrollArea.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOn)
-        self.scrollArea.setWidgetResizable(True)
-        self.scrollArea.setObjectName("scrollArea")
-        self.scrollAreaWidgetContents = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents.setGeometry(QtCore.QRect(0, 0, 591, 1000))
-        self.scrollAreaWidgetContents.setMinimumSize(QtCore.QSize(0, 1000))
-        self.scrollAreaWidgetContents.setObjectName("scrollAreaWidgetContents")
-        self.layoutWidget = QtWidgets.QWidget(self.scrollAreaWidgetContents)
-        self.layoutWidget.setGeometry(QtCore.QRect(10, 11, 523, 892))
-        self.layoutWidget.setObjectName("layoutWidget")
-        self.gridLayout = QtWidgets.QGridLayout(self.layoutWidget)
-        self.gridLayout.setContentsMargins(0, 0, 0, 0)
-        self.gridLayout.setObjectName("gridLayout")
-        self.horizontalLayout_7 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_7.setObjectName("horizontalLayout_7")
-        self.label_3 = QtWidgets.QLabel(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_3.sizePolicy().hasHeightForWidth())
-        self.label_3.setSizePolicy(sizePolicy)
-        self.label_3.setMinimumSize(QtCore.QSize(141, 31))
-        self.label_3.setMaximumSize(QtCore.QSize(141, 31))
-        self.label_3.setObjectName("label_3")
-        self.horizontalLayout_7.addWidget(self.label_3)
-        self.label_4 = QtWidgets.QLabel(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_4.sizePolicy().hasHeightForWidth())
-        self.label_4.setSizePolicy(sizePolicy)
-        self.label_4.setMinimumSize(QtCore.QSize(81, 31))
-        self.label_4.setMaximumSize(QtCore.QSize(81, 31))
-        self.label_4.setObjectName("label_4")
-        self.horizontalLayout_7.addWidget(self.label_4)
-        self.label_5 = QtWidgets.QLabel(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_5.sizePolicy().hasHeightForWidth())
-        self.label_5.setSizePolicy(sizePolicy)
-        self.label_5.setMinimumSize(QtCore.QSize(91, 31))
-        self.label_5.setMaximumSize(QtCore.QSize(100, 31))
-        self.label_5.setObjectName("label_5")
-        self.horizontalLayout_7.addWidget(self.label_5)
-        spacerItem9 = QtWidgets.QSpacerItem(110, 27, QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_7.addItem(spacerItem9)
-        self.gridLayout.addLayout(self.horizontalLayout_7, 0, 0, 1, 1)
-        self.horizontalLayout_5 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_5.setObjectName("horizontalLayout_5")
-        self.SubsysBrowser_0 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_0.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_0.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_0.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_0.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_0.setObjectName("SubsysBrowser_0")
-        self.horizontalLayout_5.addWidget(self.SubsysBrowser_0)
-        self.pktidBrowser_0 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_0.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_0.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_0.setObjectName("pktidBrowser_0")
-        self.horizontalLayout_5.addWidget(self.pktidBrowser_0)
-        self.countSpinBox_0 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_0.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_0.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_0.setReadOnly(True)
-        self.countSpinBox_0.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_0.setMaximum(16384)
-        self.countSpinBox_0.setObjectName("countSpinBox_0")
-        self.horizontalLayout_5.addWidget(self.countSpinBox_0)
-        self.pushButton_0 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_0.sizePolicy().hasHeightForWidth())
-        self.pushButton_0.setSizePolicy(sizePolicy)
-        self.pushButton_0.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_0.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_0.setObjectName("pushButton_0")
-        self.horizontalLayout_5.addWidget(self.pushButton_0)
-        self.gridLayout.addLayout(self.horizontalLayout_5, 1, 0, 1, 1)
-        self.horizontalLayout_6 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_6.setObjectName("horizontalLayout_6")
-        self.SubsysBrowser_1 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_1.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_1.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_1.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_1.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_1.setObjectName("SubsysBrowser_1")
-        self.horizontalLayout_6.addWidget(self.SubsysBrowser_1)
-        self.pktidBrowser_1 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_1.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_1.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_1.setObjectName("pktidBrowser_1")
-        self.horizontalLayout_6.addWidget(self.pktidBrowser_1)
-        self.countSpinBox_1 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_1.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_1.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_1.setReadOnly(True)
-        self.countSpinBox_1.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_1.setMaximum(16384)
-        self.countSpinBox_1.setObjectName("countSpinBox_1")
-        self.horizontalLayout_6.addWidget(self.countSpinBox_1)
-        self.pushButton_1 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_1.sizePolicy().hasHeightForWidth())
-        self.pushButton_1.setSizePolicy(sizePolicy)
-        self.pushButton_1.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_1.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_1.setObjectName("pushButton_1")
-        self.horizontalLayout_6.addWidget(self.pushButton_1)
-        self.gridLayout.addLayout(self.horizontalLayout_6, 2, 0, 1, 1)
-        self.horizontalLayout_8 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_8.setObjectName("horizontalLayout_8")
-        self.SubsysBrowser_2 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_2.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_2.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_2.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_2.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_2.setObjectName("SubsysBrowser_2")
-        self.horizontalLayout_8.addWidget(self.SubsysBrowser_2)
-        self.pktidBrowser_2 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_2.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_2.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_2.setObjectName("pktidBrowser_2")
-        self.horizontalLayout_8.addWidget(self.pktidBrowser_2)
-        self.countSpinBox_2 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_2.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_2.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_2.setReadOnly(True)
-        self.countSpinBox_2.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_2.setMaximum(16384)
-        self.countSpinBox_2.setObjectName("countSpinBox_2")
-        self.horizontalLayout_8.addWidget(self.countSpinBox_2)
-        self.pushButton_2 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_2.sizePolicy().hasHeightForWidth())
-        self.pushButton_2.setSizePolicy(sizePolicy)
-        self.pushButton_2.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_2.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_2.setObjectName("pushButton_2")
-        self.horizontalLayout_8.addWidget(self.pushButton_2)
-        self.gridLayout.addLayout(self.horizontalLayout_8, 3, 0, 1, 1)
-        self.horizontalLayout_9 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_9.setObjectName("horizontalLayout_9")
-        self.SubsysBrowser_3 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_3.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_3.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_3.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_3.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_3.setObjectName("SubsysBrowser_3")
-        self.horizontalLayout_9.addWidget(self.SubsysBrowser_3)
-        self.pktidBrowser_3 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_3.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_3.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_3.setObjectName("pktidBrowser_3")
-        self.horizontalLayout_9.addWidget(self.pktidBrowser_3)
-        self.countSpinBox_3 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_3.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_3.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_3.setReadOnly(True)
-        self.countSpinBox_3.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_3.setMaximum(16384)
-        self.countSpinBox_3.setObjectName("countSpinBox_3")
-        self.horizontalLayout_9.addWidget(self.countSpinBox_3)
-        self.pushButton_3 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_3.sizePolicy().hasHeightForWidth())
-        self.pushButton_3.setSizePolicy(sizePolicy)
-        self.pushButton_3.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_3.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_3.setObjectName("pushButton_3")
-        self.horizontalLayout_9.addWidget(self.pushButton_3)
-        self.gridLayout.addLayout(self.horizontalLayout_9, 4, 0, 1, 1)
-        self.horizontalLayout_10 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_10.setObjectName("horizontalLayout_10")
-        self.SubsysBrowser_4 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_4.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_4.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_4.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_4.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_4.setObjectName("SubsysBrowser_4")
-        self.horizontalLayout_10.addWidget(self.SubsysBrowser_4)
-        self.pktidBrowser_4 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_4.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_4.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_4.setObjectName("pktidBrowser_4")
-        self.horizontalLayout_10.addWidget(self.pktidBrowser_4)
-        self.countSpinBox_4 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_4.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_4.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_4.setReadOnly(True)
-        self.countSpinBox_4.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_4.setMaximum(16384)
-        self.countSpinBox_4.setObjectName("countSpinBox_4")
-        self.horizontalLayout_10.addWidget(self.countSpinBox_4)
-        self.pushButton_4 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_4.sizePolicy().hasHeightForWidth())
-        self.pushButton_4.setSizePolicy(sizePolicy)
-        self.pushButton_4.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_4.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_4.setObjectName("pushButton_4")
-        self.horizontalLayout_10.addWidget(self.pushButton_4)
-        self.gridLayout.addLayout(self.horizontalLayout_10, 5, 0, 1, 1)
-        self.horizontalLayout_11 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_11.setObjectName("horizontalLayout_11")
-        self.SubsysBrowser_5 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_5.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_5.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_5.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_5.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_5.setObjectName("SubsysBrowser_5")
-        self.horizontalLayout_11.addWidget(self.SubsysBrowser_5)
-        self.pktidBrowser_5 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_5.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_5.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_5.setObjectName("pktidBrowser_5")
-        self.horizontalLayout_11.addWidget(self.pktidBrowser_5)
-        self.countSpinBox_5 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_5.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_5.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_5.setReadOnly(True)
-        self.countSpinBox_5.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_5.setMaximum(16384)
-        self.countSpinBox_5.setObjectName("countSpinBox_5")
-        self.horizontalLayout_11.addWidget(self.countSpinBox_5)
-        self.pushButton_5 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_5.sizePolicy().hasHeightForWidth())
-        self.pushButton_5.setSizePolicy(sizePolicy)
-        self.pushButton_5.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_5.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_5.setObjectName("pushButton_5")
-        self.horizontalLayout_11.addWidget(self.pushButton_5)
-        self.gridLayout.addLayout(self.horizontalLayout_11, 6, 0, 1, 1)
-        self.horizontalLayout_12 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_12.setObjectName("horizontalLayout_12")
-        self.SubsysBrowser_6 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_6.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_6.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_6.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_6.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_6.setObjectName("SubsysBrowser_6")
-        self.horizontalLayout_12.addWidget(self.SubsysBrowser_6)
-        self.pktidBrowser_6 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_6.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_6.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_6.setObjectName("pktidBrowser_6")
-        self.horizontalLayout_12.addWidget(self.pktidBrowser_6)
-        self.countSpinBox_6 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_6.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_6.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_6.setReadOnly(True)
-        self.countSpinBox_6.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_6.setMaximum(16384)
-        self.countSpinBox_6.setObjectName("countSpinBox_6")
-        self.horizontalLayout_12.addWidget(self.countSpinBox_6)
-        self.pushButton_6 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_6.sizePolicy().hasHeightForWidth())
-        self.pushButton_6.setSizePolicy(sizePolicy)
-        self.pushButton_6.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_6.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_6.setObjectName("pushButton_6")
-        self.horizontalLayout_12.addWidget(self.pushButton_6)
-        self.gridLayout.addLayout(self.horizontalLayout_12, 7, 0, 1, 1)
-        self.horizontalLayout_13 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_13.setObjectName("horizontalLayout_13")
-        self.SubsysBrowser_7 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_7.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_7.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_7.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_7.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_7.setObjectName("SubsysBrowser_7")
-        self.horizontalLayout_13.addWidget(self.SubsysBrowser_7)
-        self.pktidBrowser_7 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_7.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_7.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_7.setObjectName("pktidBrowser_7")
-        self.horizontalLayout_13.addWidget(self.pktidBrowser_7)
-        self.countSpinBox_7 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_7.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_7.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_7.setReadOnly(True)
-        self.countSpinBox_7.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_7.setMaximum(16384)
-        self.countSpinBox_7.setObjectName("countSpinBox_7")
-        self.horizontalLayout_13.addWidget(self.countSpinBox_7)
-        self.pushButton_7 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_7.sizePolicy().hasHeightForWidth())
-        self.pushButton_7.setSizePolicy(sizePolicy)
-        self.pushButton_7.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_7.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_7.setObjectName("pushButton_7")
-        self.horizontalLayout_13.addWidget(self.pushButton_7)
-        self.gridLayout.addLayout(self.horizontalLayout_13, 8, 0, 1, 1)
-        self.horizontalLayout_14 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_14.setObjectName("horizontalLayout_14")
-        self.SubsysBrowser_8 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_8.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_8.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_8.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_8.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_8.setObjectName("SubsysBrowser_8")
-        self.horizontalLayout_14.addWidget(self.SubsysBrowser_8)
-        self.pktidBrowser_8 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_8.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_8.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_8.setObjectName("pktidBrowser_8")
-        self.horizontalLayout_14.addWidget(self.pktidBrowser_8)
-        self.countSpinBox_8 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_8.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_8.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_8.setReadOnly(True)
-        self.countSpinBox_8.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_8.setMaximum(16384)
-        self.countSpinBox_8.setObjectName("countSpinBox_8")
-        self.horizontalLayout_14.addWidget(self.countSpinBox_8)
-        self.pushButton_8 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_8.sizePolicy().hasHeightForWidth())
-        self.pushButton_8.setSizePolicy(sizePolicy)
-        self.pushButton_8.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_8.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_8.setObjectName("pushButton_8")
-        self.horizontalLayout_14.addWidget(self.pushButton_8)
-        self.gridLayout.addLayout(self.horizontalLayout_14, 9, 0, 1, 1)
-        self.horizontalLayout_15 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_15.setObjectName("horizontalLayout_15")
-        self.SubsysBrowser_9 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_9.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_9.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_9.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_9.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_9.setObjectName("SubsysBrowser_9")
-        self.horizontalLayout_15.addWidget(self.SubsysBrowser_9)
-        self.pktidBrowser_9 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_9.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_9.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_9.setObjectName("pktidBrowser_9")
-        self.horizontalLayout_15.addWidget(self.pktidBrowser_9)
-        self.countSpinBox_9 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_9.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_9.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_9.setReadOnly(True)
-        self.countSpinBox_9.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_9.setMaximum(16384)
-        self.countSpinBox_9.setObjectName("countSpinBox_9")
-        self.horizontalLayout_15.addWidget(self.countSpinBox_9)
-        self.pushButton_9 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_9.sizePolicy().hasHeightForWidth())
-        self.pushButton_9.setSizePolicy(sizePolicy)
-        self.pushButton_9.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_9.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_9.setObjectName("pushButton_9")
-        self.horizontalLayout_15.addWidget(self.pushButton_9)
-        self.gridLayout.addLayout(self.horizontalLayout_15, 10, 0, 1, 1)
-        self.horizontalLayout_16 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_16.setObjectName("horizontalLayout_16")
-        self.SubsysBrowser_10 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_10.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_10.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_10.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_10.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_10.setObjectName("SubsysBrowser_10")
-        self.horizontalLayout_16.addWidget(self.SubsysBrowser_10)
-        self.pktidBrowser_10 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_10.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_10.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_10.setObjectName("pktidBrowser_10")
-        self.horizontalLayout_16.addWidget(self.pktidBrowser_10)
-        self.countSpinBox_10 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_10.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_10.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_10.setReadOnly(True)
-        self.countSpinBox_10.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_10.setMaximum(16384)
-        self.countSpinBox_10.setObjectName("countSpinBox_10")
-        self.horizontalLayout_16.addWidget(self.countSpinBox_10)
-        self.pushButton_10 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_10.sizePolicy().hasHeightForWidth())
-        self.pushButton_10.setSizePolicy(sizePolicy)
-        self.pushButton_10.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_10.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_10.setObjectName("pushButton_10")
-        self.horizontalLayout_16.addWidget(self.pushButton_10)
-        self.gridLayout.addLayout(self.horizontalLayout_16, 11, 0, 1, 1)
-        self.horizontalLayout_17 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_17.setObjectName("horizontalLayout_17")
-        self.SubsysBrowser_11 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_11.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_11.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_11.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_11.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_11.setObjectName("SubsysBrowser_11")
-        self.horizontalLayout_17.addWidget(self.SubsysBrowser_11)
-        self.pktidBrowser_11 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_11.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_11.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_11.setObjectName("pktidBrowser_11")
-        self.horizontalLayout_17.addWidget(self.pktidBrowser_11)
-        self.countSpinBox_11 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_11.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_11.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_11.setReadOnly(True)
-        self.countSpinBox_11.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_11.setMaximum(16384)
-        self.countSpinBox_11.setObjectName("countSpinBox_11")
-        self.horizontalLayout_17.addWidget(self.countSpinBox_11)
-        self.pushButton_11 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_11.sizePolicy().hasHeightForWidth())
-        self.pushButton_11.setSizePolicy(sizePolicy)
-        self.pushButton_11.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_11.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_11.setObjectName("pushButton_11")
-        self.horizontalLayout_17.addWidget(self.pushButton_11)
-        self.gridLayout.addLayout(self.horizontalLayout_17, 12, 0, 1, 1)
-        self.horizontalLayout_18 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_18.setObjectName("horizontalLayout_18")
-        self.SubsysBrowser_12 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_12.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_12.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_12.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_12.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_12.setObjectName("SubsysBrowser_12")
-        self.horizontalLayout_18.addWidget(self.SubsysBrowser_12)
-        self.pktidBrowser_12 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_12.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_12.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_12.setObjectName("pktidBrowser_12")
-        self.horizontalLayout_18.addWidget(self.pktidBrowser_12)
-        self.countSpinBox_12 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_12.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_12.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_12.setReadOnly(True)
-        self.countSpinBox_12.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_12.setMaximum(16384)
-        self.countSpinBox_12.setObjectName("countSpinBox_12")
-        self.horizontalLayout_18.addWidget(self.countSpinBox_12)
-        self.pushButton_12 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_12.sizePolicy().hasHeightForWidth())
-        self.pushButton_12.setSizePolicy(sizePolicy)
-        self.pushButton_12.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_12.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_12.setObjectName("pushButton_12")
-        self.horizontalLayout_18.addWidget(self.pushButton_12)
-        self.gridLayout.addLayout(self.horizontalLayout_18, 13, 0, 1, 1)
-        self.horizontalLayout_19 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_19.setObjectName("horizontalLayout_19")
-        self.SubsysBrowser_13 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_13.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_13.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_13.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_13.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_13.setObjectName("SubsysBrowser_13")
-        self.horizontalLayout_19.addWidget(self.SubsysBrowser_13)
-        self.pktidBrowser_13 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_13.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_13.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_13.setObjectName("pktidBrowser_13")
-        self.horizontalLayout_19.addWidget(self.pktidBrowser_13)
-        self.countSpinBox_13 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_13.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_13.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_13.setReadOnly(True)
-        self.countSpinBox_13.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_13.setMaximum(16384)
-        self.countSpinBox_13.setObjectName("countSpinBox_13")
-        self.horizontalLayout_19.addWidget(self.countSpinBox_13)
-        self.pushButton_13 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_13.sizePolicy().hasHeightForWidth())
-        self.pushButton_13.setSizePolicy(sizePolicy)
-        self.pushButton_13.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_13.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_13.setObjectName("pushButton_13")
-        self.horizontalLayout_19.addWidget(self.pushButton_13)
-        self.gridLayout.addLayout(self.horizontalLayout_19, 14, 0, 1, 1)
-        self.horizontalLayout_20 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_20.setObjectName("horizontalLayout_20")
-        self.SubsysBrowser_14 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_14.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_14.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_14.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_14.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_14.setObjectName("SubsysBrowser_14")
-        self.horizontalLayout_20.addWidget(self.SubsysBrowser_14)
-        self.pktidBrowser_14 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_14.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_14.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_14.setObjectName("pktidBrowser_14")
-        self.horizontalLayout_20.addWidget(self.pktidBrowser_14)
-        self.countSpinBox_14 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_14.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_14.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_14.setReadOnly(True)
-        self.countSpinBox_14.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_14.setMaximum(16384)
-        self.countSpinBox_14.setObjectName("countSpinBox_14")
-        self.horizontalLayout_20.addWidget(self.countSpinBox_14)
-        self.pushButton_14 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_14.sizePolicy().hasHeightForWidth())
-        self.pushButton_14.setSizePolicy(sizePolicy)
-        self.pushButton_14.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_14.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_14.setObjectName("pushButton_14")
-        self.horizontalLayout_20.addWidget(self.pushButton_14)
-        self.gridLayout.addLayout(self.horizontalLayout_20, 15, 0, 1, 1)
-        self.horizontalLayout_21 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_21.setObjectName("horizontalLayout_21")
-        self.SubsysBrowser_15 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_15.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_15.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_15.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_15.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_15.setObjectName("SubsysBrowser_15")
-        self.horizontalLayout_21.addWidget(self.SubsysBrowser_15)
-        self.pktidBrowser_15 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_15.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_15.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_15.setObjectName("pktidBrowser_15")
-        self.horizontalLayout_21.addWidget(self.pktidBrowser_15)
-        self.countSpinBox_15 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_15.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_15.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_15.setReadOnly(True)
-        self.countSpinBox_15.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_15.setMaximum(16384)
-        self.countSpinBox_15.setObjectName("countSpinBox_15")
-        self.horizontalLayout_21.addWidget(self.countSpinBox_15)
-        self.pushButton_15 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_15.sizePolicy().hasHeightForWidth())
-        self.pushButton_15.setSizePolicy(sizePolicy)
-        self.pushButton_15.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_15.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_15.setObjectName("pushButton_15")
-        self.horizontalLayout_21.addWidget(self.pushButton_15)
-        self.gridLayout.addLayout(self.horizontalLayout_21, 16, 0, 1, 1)
-        self.horizontalLayout_22 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_22.setObjectName("horizontalLayout_22")
-        self.SubsysBrowser_16 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_16.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_16.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_16.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_16.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_16.setObjectName("SubsysBrowser_16")
-        self.horizontalLayout_22.addWidget(self.SubsysBrowser_16)
-        self.pktidBrowser_16 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_16.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_16.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_16.setObjectName("pktidBrowser_16")
-        self.horizontalLayout_22.addWidget(self.pktidBrowser_16)
-        self.countSpinBox_16 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_16.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_16.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_16.setReadOnly(True)
-        self.countSpinBox_16.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_16.setMaximum(16384)
-        self.countSpinBox_16.setObjectName("countSpinBox_16")
-        self.horizontalLayout_22.addWidget(self.countSpinBox_16)
-        self.pushButton_16 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_16.sizePolicy().hasHeightForWidth())
-        self.pushButton_16.setSizePolicy(sizePolicy)
-        self.pushButton_16.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_16.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_16.setObjectName("pushButton_16")
-        self.horizontalLayout_22.addWidget(self.pushButton_16)
-        self.gridLayout.addLayout(self.horizontalLayout_22, 17, 0, 1, 1)
-        self.horizontalLayout_23 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_23.setObjectName("horizontalLayout_23")
-        self.SubsysBrowser_17 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_17.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_17.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_17.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_17.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_17.setObjectName("SubsysBrowser_17")
-        self.horizontalLayout_23.addWidget(self.SubsysBrowser_17)
-        self.pktidBrowser_17 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_17.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_17.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_17.setObjectName("pktidBrowser_17")
-        self.horizontalLayout_23.addWidget(self.pktidBrowser_17)
-        self.countSpinBox_17 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_17.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_17.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_17.setReadOnly(True)
-        self.countSpinBox_17.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_17.setMaximum(16384)
-        self.countSpinBox_17.setObjectName("countSpinBox_17")
-        self.horizontalLayout_23.addWidget(self.countSpinBox_17)
-        self.pushButton_17 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_17.sizePolicy().hasHeightForWidth())
-        self.pushButton_17.setSizePolicy(sizePolicy)
-        self.pushButton_17.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_17.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_17.setObjectName("pushButton_17")
-        self.horizontalLayout_23.addWidget(self.pushButton_17)
-        self.gridLayout.addLayout(self.horizontalLayout_23, 18, 0, 1, 1)
-        self.horizontalLayout_24 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_24.setObjectName("horizontalLayout_24")
-        self.SubsysBrowser_18 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_18.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_18.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_18.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_18.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_18.setObjectName("SubsysBrowser_18")
-        self.horizontalLayout_24.addWidget(self.SubsysBrowser_18)
-        self.pktidBrowser_18 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_18.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_18.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_18.setObjectName("pktidBrowser_18")
-        self.horizontalLayout_24.addWidget(self.pktidBrowser_18)
-        self.countSpinBox_18 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_18.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_18.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_18.setReadOnly(True)
-        self.countSpinBox_18.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_18.setMaximum(16384)
-        self.countSpinBox_18.setObjectName("countSpinBox_18")
-        self.horizontalLayout_24.addWidget(self.countSpinBox_18)
-        self.pushButton_18 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_18.sizePolicy().hasHeightForWidth())
-        self.pushButton_18.setSizePolicy(sizePolicy)
-        self.pushButton_18.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_18.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_18.setObjectName("pushButton_18")
-        self.horizontalLayout_24.addWidget(self.pushButton_18)
-        self.gridLayout.addLayout(self.horizontalLayout_24, 19, 0, 1, 1)
-        self.horizontalLayout_25 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_25.setObjectName("horizontalLayout_25")
-        self.SubsysBrowser_19 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_19.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_19.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_19.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_19.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_19.setObjectName("SubsysBrowser_19")
-        self.horizontalLayout_25.addWidget(self.SubsysBrowser_19)
-        self.pktidBrowser_19 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_19.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_19.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_19.setObjectName("pktidBrowser_19")
-        self.horizontalLayout_25.addWidget(self.pktidBrowser_19)
-        self.countSpinBox_19 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_19.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_19.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_19.setReadOnly(True)
-        self.countSpinBox_19.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_19.setMaximum(16384)
-        self.countSpinBox_19.setObjectName("countSpinBox_19")
-        self.horizontalLayout_25.addWidget(self.countSpinBox_19)
-        self.pushButton_19 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_19.sizePolicy().hasHeightForWidth())
-        self.pushButton_19.setSizePolicy(sizePolicy)
-        self.pushButton_19.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_19.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_19.setObjectName("pushButton_19")
-        self.horizontalLayout_25.addWidget(self.pushButton_19)
-        self.gridLayout.addLayout(self.horizontalLayout_25, 20, 0, 1, 1)
-        self.horizontalLayout_26 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_26.setObjectName("horizontalLayout_26")
-        self.SubsysBrowser_20 = QtWidgets.QPlainTextEdit(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.SubsysBrowser_20.sizePolicy().hasHeightForWidth())
-        self.SubsysBrowser_20.setSizePolicy(sizePolicy)
-        self.SubsysBrowser_20.setMinimumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_20.setMaximumSize(QtCore.QSize(141, 31))
-        self.SubsysBrowser_20.setObjectName("SubsysBrowser_20")
-        self.horizontalLayout_26.addWidget(self.SubsysBrowser_20)
-        self.pktidBrowser_20 = QtWidgets.QLineEdit(self.layoutWidget)
-        self.pktidBrowser_20.setMinimumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_20.setMaximumSize(QtCore.QSize(81, 31))
-        self.pktidBrowser_20.setObjectName("pktidBrowser_20")
-        self.horizontalLayout_26.addWidget(self.pktidBrowser_20)
-        self.countSpinBox_20 = QtWidgets.QSpinBox(self.layoutWidget)
-        self.countSpinBox_20.setMinimumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_20.setMaximumSize(QtCore.QSize(81, 31))
-        self.countSpinBox_20.setReadOnly(True)
-        self.countSpinBox_20.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
-        self.countSpinBox_20.setMaximum(16384)
-        self.countSpinBox_20.setObjectName("countSpinBox_20")
-        self.horizontalLayout_26.addWidget(self.countSpinBox_20)
-        self.pushButton_20 = QtWidgets.QPushButton(self.layoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pushButton_20.sizePolicy().hasHeightForWidth())
-        self.pushButton_20.setSizePolicy(sizePolicy)
-        self.pushButton_20.setMinimumSize(QtCore.QSize(101, 27))
-        self.pushButton_20.setMaximumSize(QtCore.QSize(120, 27))
-        self.pushButton_20.setObjectName("pushButton_20")
-        self.horizontalLayout_26.addWidget(self.pushButton_20)
-        self.gridLayout.addLayout(self.horizontalLayout_26, 21, 0, 1, 1)
-        self.scrollArea.setWidget(self.scrollAreaWidgetContents)
-        self.verticalLayout.addWidget(self.scrollArea)
-
-        self.retranslateUi(TelemetrySystemDialog)
-        self.buttonBox.clicked['QAbstractButton*'].connect(TelemetrySystemDialog.close)
-        QtCore.QMetaObject.connectSlotsByName(TelemetrySystemDialog)
-
-    def retranslateUi(self, TelemetrySystemDialog):
-        _translate = QtCore.QCoreApplication.translate
-        TelemetrySystemDialog.setWindowTitle(_translate("TelemetrySystemDialog", "Dialog"))
-        self.label.setText(_translate("TelemetrySystemDialog", "cFE/CFS Subsystem Telemetry"))
-        self.label_6.setText(_translate("TelemetrySystemDialog", "Packets Received"))
-        self.label_2.setText(_translate("TelemetrySystemDialog", "Available Pages"))
-        self.label_3.setText(_translate("TelemetrySystemDialog", "Subsystem/Page"))
-        self.label_4.setText(_translate("TelemetrySystemDialog", "Packet ID"))
-        self.label_5.setText(_translate("TelemetrySystemDialog", "Packet Count"))
-        self.pushButton_0.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_1.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_2.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_3.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_4.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_5.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_6.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_7.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_8.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_9.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_10.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_11.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_12.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_13.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_14.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_15.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_16.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_17.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_18.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_19.setText(_translate("TelemetrySystemDialog", "Display Page"))
-        self.pushButton_20.setText(_translate("TelemetrySystemDialog", "Display Page"))
-
