+++ Subsystems/tlmGUI/Ui_GenericTelemetryDialog.py
@@ -1,0 +1,110 @@
+# -*- coding: utf-8 -*-
+
+# Form implementation generated from reading ui file '/home/lbleier/cFS/tools/cFS-GroundSystem/Subsystems/tlmGUI/GenericTelemetryDialog.ui'
+#
+# Created by: PyQt5 UI code generator 5.10.1
+#
+# WARNING! All changes made in this file will be lost!
+
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+class Ui_GenericTelemetryDialog(object):
+    def setupUi(self, GenericTelemetryDialog):
+        GenericTelemetryDialog.setObjectName("GenericTelemetryDialog")
+        GenericTelemetryDialog.resize(630, 816)
+        self.label_6 = QtWidgets.QLabel(GenericTelemetryDialog)
+        self.label_6.setGeometry(QtCore.QRect(90, 730, 431, 61))
+        font = QtGui.QFont()
+        font.setPointSize(12)
+        self.label_6.setFont(font)
+        self.label_6.setObjectName("label_6")
+        self.tblTelemetry = QtWidgets.QTableWidget(GenericTelemetryDialog)
+        self.tblTelemetry.setGeometry(QtCore.QRect(70, 110, 470, 601))
+        self.tblTelemetry.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOff)
+        self.tblTelemetry.setSizeAdjustPolicy(QtWidgets.QAbstractScrollArea.AdjustToContentsOnFirstShow)
+        self.tblTelemetry.setEditTriggers(QtWidgets.QAbstractItemView.NoEditTriggers)
+        self.tblTelemetry.setObjectName("tblTelemetry")
+        self.tblTelemetry.setColumnCount(2)
+        self.tblTelemetry.setRowCount(0)
+        item = QtWidgets.QTableWidgetItem()
+        self.tblTelemetry.setHorizontalHeaderItem(0, item)
+        item = QtWidgets.QTableWidgetItem()
+        self.tblTelemetry.setHorizontalHeaderItem(1, item)
+        self.tblTelemetry.horizontalHeader().setDefaultSectionSize(225)
+        self.buttonBox = QtWidgets.QDialogButtonBox(GenericTelemetryDialog)
+        self.buttonBox.setGeometry(QtCore.QRect(510, 50, 81, 31))
+        self.buttonBox.setStandardButtons(QtWidgets.QDialogButtonBox.Close)
+        self.buttonBox.setObjectName("buttonBox")
+        self.widget = QtWidgets.QWidget(GenericTelemetryDialog)
+        self.widget.setGeometry(QtCore.QRect(270, 25, 73, 60))
+        self.widget.setObjectName("widget")
+        self.verticalLayout = QtWidgets.QVBoxLayout(self.widget)
+        self.verticalLayout.setContentsMargins(0, 0, 0, 0)
+        self.verticalLayout.setObjectName("verticalLayout")
+        self.packetIdLabel = QtWidgets.QLabel(self.widget)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.packetIdLabel.sizePolicy().hasHeightForWidth())
+        self.packetIdLabel.setSizePolicy(sizePolicy)
+        self.packetIdLabel.setMinimumSize(QtCore.QSize(0, 21))
+        self.packetIdLabel.setMaximumSize(QtCore.QSize(16777215, 25))
+        self.packetIdLabel.setObjectName("packetIdLabel")
+        self.verticalLayout.addWidget(self.packetIdLabel)
+        self.packetId = QtWidgets.QLCDNumber(self.widget)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Minimum)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.packetId.sizePolicy().hasHeightForWidth())
+        self.packetId.setSizePolicy(sizePolicy)
+        self.packetId.setObjectName("packetId")
+        self.verticalLayout.addWidget(self.packetId)
+        self.widget1 = QtWidgets.QWidget(GenericTelemetryDialog)
+        self.widget1.setGeometry(QtCore.QRect(370, 23, 125, 62))
+        self.widget1.setObjectName("widget1")
+        self.verticalLayout_3 = QtWidgets.QVBoxLayout(self.widget1)
+        self.verticalLayout_3.setContentsMargins(0, 0, 0, 0)
+        self.verticalLayout_3.setObjectName("verticalLayout_3")
+        self.label_5 = QtWidgets.QLabel(self.widget1)
+        self.label_5.setObjectName("label_5")
+        self.verticalLayout_3.addWidget(self.label_5)
+        self.sequenceCount = QtWidgets.QSpinBox(self.widget1)
+        self.sequenceCount.setMinimumSize(QtCore.QSize(101, 31))
+        self.sequenceCount.setMaximumSize(QtCore.QSize(101, 31))
+        self.sequenceCount.setButtonSymbols(QtWidgets.QAbstractSpinBox.NoButtons)
+        self.sequenceCount.setMaximum(16384)
+        self.sequenceCount.setProperty("value", 0)
+        self.sequenceCount.setObjectName("sequenceCount")
+        self.verticalLayout_3.addWidget(self.sequenceCount)
+        self.widget2 = QtWidgets.QWidget(GenericTelemetryDialog)
+        self.widget2.setGeometry(QtCore.QRect(42, 24, 204, 62))
+        self.widget2.setObjectName("widget2")
+        self.verticalLayout_2 = QtWidgets.QVBoxLayout(self.widget2)
+        self.verticalLayout_2.setContentsMargins(0, 0, 0, 0)
+        self.verticalLayout_2.setObjectName("verticalLayout_2")
+        self.subSystemTelemetryPageLabel = QtWidgets.QLabel(self.widget2)
+        self.subSystemTelemetryPageLabel.setObjectName("subSystemTelemetryPageLabel")
+        self.verticalLayout_2.addWidget(self.subSystemTelemetryPageLabel)
+        self.subSystemLineEdit = QtWidgets.QLineEdit(self.widget2)
+        self.subSystemLineEdit.setMinimumSize(QtCore.QSize(141, 31))
+        self.subSystemLineEdit.setMaximumSize(QtCore.QSize(141, 31))
+        self.subSystemLineEdit.setReadOnly(True)
+        self.subSystemLineEdit.setObjectName("subSystemLineEdit")
+        self.verticalLayout_2.addWidget(self.subSystemLineEdit)
+
+        self.retranslateUi(GenericTelemetryDialog)
+        self.buttonBox.clicked['QAbstractButton*'].connect(GenericTelemetryDialog.close)
+        QtCore.QMetaObject.connectSlotsByName(GenericTelemetryDialog)
+
+    def retranslateUi(self, GenericTelemetryDialog):
+        _translate = QtCore.QCoreApplication.translate
+        GenericTelemetryDialog.setWindowTitle(_translate("GenericTelemetryDialog", "Telemetry Page"))
+        self.label_6.setText(_translate("GenericTelemetryDialog", "<html><head/><body><p>*No packets? Remember to select the IP address of your<br/>spacecraft in the Main Window.</p></body></html>"))
+        item = self.tblTelemetry.horizontalHeaderItem(0)
+        item.setText(_translate("GenericTelemetryDialog", "Telemetry Point Label"))
+        item = self.tblTelemetry.horizontalHeaderItem(1)
+        item.setText(_translate("GenericTelemetryDialog", "Telemetry Point Value"))
+        self.packetIdLabel.setText(_translate("GenericTelemetryDialog", "Packet ID"))
+        self.label_5.setText(_translate("GenericTelemetryDialog", "Sequence Count"))
+        self.subSystemTelemetryPageLabel.setText(_translate("GenericTelemetryDialog", "Subsystem Telemetry Page"))
+
